Module uom::marker

Primitive traits and types representing basic properties of types.

Traits

Add
Trait to denote that a quantity is able to be added with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Add is implemented automatically.

AddAssign
Trait to denote that a quantity is able to be added with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::AddAssign is implemented automatically.

Div
Trait to denote that a quantity is able to be divided with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Div is implemented automatically.

DivAssign
Trait to denote that a quantity is able to be divided with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::DivAssign is implemented automatically.

Mul
Trait to denote that a quantity is able to be multiplied with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Mul is implemented automatically.

MulAssign
Trait to denote that a quantity is able to be multiplied with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::MulAssign is implemented automatically.

Neg
Trait to denote that a quantity is able to be negated. When a specific quantity's kind inherits this trait ops::Neg is implemented automatically.

Rem
Trait to denote that a quantity is able to calculate a remainder with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Rem is implemented automatically.

RemAssign
Trait to denote that a quantity is able to calculate a remainder with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::RemAssign is implemented automatically.

Saturating
Trait to denote that a quantity is able to perform saturating additions and subtractions with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Saturating is implemented automatically.

Sub
Trait to denote that a quantity is able to be subtracted with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::Sub is implemented automatically.

SubAssign
Trait to denote that a quantity is able to be subtracted with a quantity of the same dimensions. When a specific quantity's kind inherits this trait ops::SubAssign is implemented automatically.