Module uom::si::amount_of_substance

Amount of substance (base unit mole, mol).

Structs

attomole
Measurement unit.

centimole
Measurement unit.

decamole
Measurement unit.

decimole
Measurement unit.

examole
Measurement unit.

femtomole
Measurement unit.

gigamole
Measurement unit.

hectomole
Measurement unit.

kilomole
Measurement unit.

megamole
Measurement unit.

micromole
Measurement unit.

millimole
Measurement unit.

mole

The mole is the SI unit of amount of substance. One mole contains exactly 6.022 140 76 × 10²³ elementary entities. This number is the fixed numerical value of the Avogadro constant, NA, when expressed in the unit mol⁻¹ and is called the Avogadro number.

The amount of substance, symbol n, of a system is a measure of the number of specified elementary entities. An elementary entity may be an atom, a molecule, an ion, an electron, any other particle or specified group of particles.

nanomole
Measurement unit.

petamole
Measurement unit.

picomole
Measurement unit.

teramole
Measurement unit.

yoctomole
Measurement unit.

yottamole
Measurement unit.

zeptomole
Measurement unit.

zettamole
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

AmountOfSubstance
Amount of substance (base unit mole, mol).

Dimension
Dimension of amount of substance, N (base unit mole, mol).