Module uom::si::angle

Angle (dimensionless quantity).

Structs

degree
Measurement unit.

gon
Measurement unit.

mil
Measurement unit.

minute
Measurement unit.

radian
SI derived unit of angle. It is the angle subtended at the center of a circle by an arc that is equal in length to the radius of the circle.

revolution
Measurement unit.

second
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Angle
Angle (dimensionless quantity).

Dimension
Dimension of angle, 1 (dimensionless).