Module uom::si::angular_velocity

Angular velocity (base unit radian per second, s⁻¹).

Structs

degree_per_second
Measurement unit.

radian_per_second
Derived unit of angular velocity.

revolution_per_hour
Measurement unit.

revolution_per_minute
Measurement unit.

revolution_per_second
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

AngularVelocity
Angular velocity (base unit radian per second, s⁻¹).

Dimension
Dimension of angular velocity, T⁻¹ (base unit radian per second, s⁻¹).