Module uom::si::areal_number_density

Areal number density (base unit 1 per square meter, m⁻²).

Structs

per_acre

per acre

per_are

per are

per_barn

per barn

per_circular_mil

per circular mil

per_hectare

per hectare

per_square_centimeter

per square centimeter

per_square_decimeter

per square decimeter

per_square_foot

per square foot

per_square_inch

per square inch

per_square_kilometer

per square kilometer

per_square_meter

per square meter

per_square_micrometer

per square micrometer

per_square_mile

per square mile

per_square_millimeter

per square millimeter

per_square_yard

per square yard

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

ArealNumberDensity

Areal number density (base unit 1 per square meter, m⁻²).

Dimension

Dimension of areal number density, L⁻² (base unit 1 per square meter, m⁻²).