Module uom::si::capacitance

Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

Structs

abfarad
Measurement unit.

attofarad
Measurement unit.

centifarad
Measurement unit.

decafarad
Measurement unit.

decifarad
Measurement unit.

exafarad
Measurement unit.

farad
Derived unit of capacitance.

femtofarad
Measurement unit.

gigafarad
Measurement unit.

hectofarad
Measurement unit.

kilofarad
Measurement unit.

megafarad
Measurement unit.

microfarad
Measurement unit.

millifarad
Measurement unit.

nanofarad
Measurement unit.

petafarad
Measurement unit.

picofarad
Measurement unit.

statfarad
Measurement unit.

terafarad
Measurement unit.

yoctofarad
Measurement unit.

yottafarad
Measurement unit.

zeptofarad
Measurement unit.

zettafarad
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Capacitance
Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

Dimension
Dimension of capacitance, L⁻²M⁻¹T⁴I² (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).