Module uom::si::electrical_conductance

Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

Structs

abmho

abmhos

absiemens

abmsiemens

attosiemens

attosiemens

centisiemens

centisiemens

decasiemens

decasiemens

decisiemens

decisiemens

exasiemens

exasiemens

femtosiemens

femtosiemens

gigasiemens

gigasiemens

hectosiemens

hectosiemens

kilosiemens

kilosiemens

megasiemens

megasiemens

mho

mhos

microsiemens

microsiemens

millisiemens

millisiemens

nanosiemens

nanosiemens

petasiemens

petasiemens

picosiemens

picosiemens

siemens

Derived unit of electrical conductance.

statmho

statmhos

statsiemens

statsiemens

terasiemens

terasiemens

yoctosiemens

yoctosiemens

yottasiemens

yottasiemens

zeptosiemens

zeptosiemens

zettasiemens

zettasiemens

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of electrical conductance, L⁻²M⁻¹T³I² (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

ElectricalConductance

Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).