Module uom::si::electrical_resistance

Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).

Structs

abohm
Measurement unit.

attoohm
Measurement unit.

centiohm
Measurement unit.

decaohm
Measurement unit.

deciohm
Measurement unit.

exaohm
Measurement unit.

femtoohm
Measurement unit.

gigaohm
Measurement unit.

hectoohm
Measurement unit.

kiloohm
Measurement unit.

megaohm
Measurement unit.

microohm
Measurement unit.

milliohm
Measurement unit.

nanoohm
Measurement unit.

ohm
Derived unit of electrical resistance.

petaohm
Measurement unit.

picoohm
Measurement unit.

statohm
Measurement unit.

teraohm
Measurement unit.

yoctoohm
Measurement unit.

yottaohm
Measurement unit.

zeptoohm
Measurement unit.

zettaohm
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of electrical resistance, L²MT⁻³I⁻² (base unit ohm, m² · kg · s⁻³ · A⁻²).

ElectricalResistance
Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).