Module uom::si::energy

Energy (base unit joule, kg · m² · s⁻²).

Structs

attojoule

attojoules

btu

British thermal units

btu_39

British thermal units (39 °F)

btu_59

British thermal units (59 °F)

btu_60

British thermal units (60 °F)

btu_it

British thermal units (IT)

calorie

calories

calorie_15

calories (15 °C)

calorie_20

calories (20 °C)

calorie_it

calories (IT)

calorie_it_nutrition

Calories (IT)

calorie_nutrition

Calories

centijoule

centijoules

decaelectronvolt

decaelectronvolts

decajoule

decajoules

decawatt_hour

decawatt hours

decijoule

decijoules

electronvolt

electronvolts

erg

ergs

exajoule

exajoules

femtojoule

femtojoules

foot_pound

foot pounds-force

foot_poundal

foot poundals

gigaelectronvolt

gigaelectronvolts

gigajoule

gigajoules

gigawatt_hour

gigawatt hours

hartree

Atomic unit of energy (Hartree energy).

hectoelectronvolt

hectoelectronvolts

hectojoule

hectojoules

hectowatt_hour

hectowatt hours

joule

Derived unit of energy.

kilocalorie

kilocalories

kilocalorie_it

kilocalories (IT)

kiloelectronvolt

kiloelectronvolts

kilojoule

kilojoules

kilowatt_hour

kilowatt hours

megaelectronvolt

megaelectronvolts

megajoule

megajoules

megawatt_hour

megawatt hours

microjoule

microjoules

microwatt_hour

microwatt hours

millijoule

millijoules

milliwatt_hour

milliwatt hours

nanojoule

nanojoules

petaelectronvolt

petaelectronvolts

petajoule

petajoules

petawatt_hour

petawatt hours

picojoule

picojoules

quad

quads

teraelectronvolt

teraelectronvolts

terajoule

terajoules

terawatt_hour

terawatt hours

therm_ec

therms (EC)

therm_us

therms

ton_tnt

tons of TNT

watt_hour

watt hours

watt_second

watt seconds

yoctojoule

yoctojoules

yottajoule

yottajoules

zeptojoule

zeptojoules

zettajoule

zettajoules

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of energy, L²MT⁻² (base unit joule, kg · m² · s⁻²).

Energy

Energy (base unit joule, kg · m² · s⁻²).