Module uom::si::force

Force (base unit newton, kg · m · s⁻²).

Structs

attonewton

attonewtons

centinewton

centinewtons

decanewton

decanewtons

decinewton

decinewtons

dyne

dynes

exanewton

exanewtons

femtonewton

femtonewtons

giganewton

giganewtons

hectonewton

hectonewtons

kilogram_force

kilograms-force

kilonewton

kilonewtons

kip

kips

meganewton

meganewtons

micronewton

micronewtons

millinewton

millinewtons

nanonewton

nanonewtons

newton

Derived unit of force.

ounce_force

ounces-force

petanewton

petanewtons

piconewton

piconewtons

pound_force

pounds-force

poundal

poundals

teranewton

teranewtons

ton_force

tons-force

yoctonewton

yoctonewtons

yottanewton

yottanewtons

zeptonewton

zeptonewtons

zettanewton

zettanewtons

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of force, LMT⁻² (base unit newton, kg · m · s⁻²).

Force

Force (base unit newton, kg · m · s⁻²).