Module uom::si::heat_flux_density

Heat flux density (base unit watt per square meter, kg · s⁻³).

Structs

attowatt_per_square_meter
Measurement unit.

centiwatt_per_square_meter
Measurement unit.

decawatt_per_square_meter
Measurement unit.

deciwatt_per_square_meter
Measurement unit.

exawatt_per_square_meter
Measurement unit.

femtowatt_per_square_meter
Measurement unit.

gigawatt_per_square_meter
Measurement unit.

hectowatt_per_square_meter
Measurement unit.

kilowatt_per_square_meter
Derived unit of density.

megawatt_per_square_meter
Measurement unit.

microwatt_per_square_meter
Measurement unit.

milliwatt_per_square_meter
Measurement unit.

nanowatt_per_square_meter
Measurement unit.

petawatt_per_square_meter
Measurement unit.

picowatt_per_square_meter
Measurement unit.

terawatt_per_square_meter
Measurement unit.

watt_per_square_meter
Measurement unit.

yoctowatt_per_square_meter
Measurement unit.

yottawatt_per_square_meter
Measurement unit.

zeptowatt_per_square_meter
Measurement unit.

zettawatt_per_square_meter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of heat flux density, MT⁻³ (base unit watt per square meter, kg · s⁻³).

HeatFluxDensity
Heat flux density (base unit watt per square meter, kg · s⁻³).