Module uom::si::information

Information (dimensionless quantity).

Structs

bit

The bit is the basic unit of information.

byte

The byte is a unit of information consisting of 8 bits.

crumb

crumbs

deciban

decibans

exabit

exabits

exabyte

exabytes

exbibit

exbibits

exbibyte

exbibytes

gibibit

gibibits

gibibyte

gibibytes

gigabit

gigabits

gigabyte

gigabytes

hartley

hartleys

kibibit

kibibits

kibibyte

kibibytes

kilobit

kilobits

kilobyte

kilobytes

mebibit

mebibits

mebibyte

mebibytes

megabit

megabits

megabyte

megabytes

natural_unit_of_information

natural units of information

nibble

nibbles

octet

octets

pebibit

pebibits

pebibyte

pebibytes

petabit

petabits

petabyte

petabytes

shannon

shannons

tebibit

tibibits

tebibyte

tibibytes

terabit

terabits

terabyte

terabytes

trit

trits

yobibit

yobibits

yobibyte

yobibytes

yottabit

yottabits

yottabyte

yottabytes

zebibit

zebibits

zebibyte

zebibytes

zettabit

zettabits

zettabyte

zettabytes

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of information, 1 (dimensionless).

Information

Information (dimensionless quantity).