Module uom::si::linear_number_rate

Linear number rate (base unit 1 per meter second, m⁻¹ · s⁻¹).

Structs

per_centimeter_second

per centimeter second

per_foot_second

per foot second

per_foot_survey_second

per foot (U.S. survey) second

per_inch_second

per inch second

per_kilometer_second

per kilometer second

per_meter_second

per meter second

per_mile_second

per mile second

per_mile_survey_second

per mile (U.S. survey) second

per_millimeter_second

per millimeter second

per_yard_second

per yard second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of linear number rate, L⁻¹T⁻¹ (base unit 1 per meter second, m⁻¹ · s⁻¹).

LinearNumberRate

Linear number rate (base unit 1 per meter second, m⁻¹ · s⁻¹).