Module uom::si::luminous_intensity

Luminous intensity (base unit candela, cd).

Structs

attocandela

attocandelas

candela

The candela is the SI unit of luminous intensity in a given direction. It is defined by taking the fixed numerical value of the luminous efficacy of monochromatic radiation of frequency 540 × 10¹² Hz, Kcd, to be 683 when expressed in the unit lm W⁻¹, which is equal to cd sr W⁻¹, or cd sr kg⁻¹ m⁻² s³, where the kilogram, meter, and second are defined in terms of h, c and ∆νCs.

centicandela

centicandelas

decacandela

decacandelas

decicandela

decicandelas

exacandela

exacandelas

femtocandela

femtocandelas

gigacandela

gigacandelas

hectocandela

hectocandelas

kilocandela

kilocandelas

megacandela

megacandelas

microcandela

microcandelas

millicandela

millicandelas

nanocandela

nanocandelas

petacandela

petacandelas

picocandela

picocandelas

teracandela

teracandelas

yoctocandela

yoctocandelas

yottacandela

yottacandelas

zeptocandela

zeptocandelas

zettacandela

zettacandelas

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of luminous intensity, J (base unit candela, cd).

LuminousIntensity

Luminous intensity (base unit candela, cd).