Module uom::si::magnetic_flux

Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).

Structs

attoweber

attowebers

centiweber

centiwebers

decaweber

decawebers

deciweber

deciwebers

exaweber

exawebers

femtoweber

femtowebers

gigaweber

gigawebers

hectoweber

hectowebers

kiloweber

kilowebers

maxwell

maxwells

megaweber

megawebers

microweber

microwebers

milliweber

milliwebers

nanoweber

nanowebers

petaweber

petawebers

picoweber

picowebers

teraweber

terawebers

weber

Derived unit of magnetic flux.

yoctoweber

yoctowebers

yottaweber

yottawebers

zeptoweber

zeptowebers

zettaweber

zettawebers

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of magnetic flux, L²MT⁻²I⁻¹ (base unit weber, m² · kg · s⁻² · A⁻¹).

MagneticFlux

Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).