Module uom::si::magnetic_flux_density

Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).

Structs

attotesla

attoteslas

centitesla

centiteslas

decatesla

decateslas

decitesla

deciteslas

exatesla

exateslas

femtotesla

femtoteslas

gamma

gammas

gauss

gauss

gigatesla

gigateslas

hectotesla

hectoteslas

kilotesla

kiloteslas

megatesla

megateslas

microtesla

microteslas

millitesla

milliteslas

nanotesla

nanoteslas

petatesla

petateslas

picotesla

picoteslas

teratesla

terateslas

tesla

Derived unit of magnetic flux density.

yoctotesla

yoctoteslas

yottatesla

yottateslas

zeptotesla

zeptoteslas

zettatesla

zettateslas

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of magnetic flux density, MT⁻²I⁻¹ (base unit tesla, kg · s⁻² · A⁻¹).

MagneticFluxDensity

Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).