Module uom::si::magnetic_moment

Magnetic moment (base unit ampere square meter A · m²).

Structs

ampere_square_centimeter

ampere square centimeters

ampere_square_meter

ampere square meters

atomic_unit_of_magnetic_dipole_moment

atomic units of magnetic dipole moment

bohr_magneton

Bohr magnetons

deuteron_magnetic_moment

deuteron magnetic moments

electron_magnetic_moment

electron magnetic moments

erg_per_gauss

ergs per gauss

helion_magnetic_moment

helion magnetic moments

joule_per_tesla

joules per tesla

muon_magnetic_moment

muon magnetic moments

neutron_magnetic_moment

neutron magnetic moments

newton_meter_per_tesla

newton meters per tesla

nuclear_magneton

nuclear magnetons

proton_magnetic_moment

proton magnetic moments

shielded_helion_magnetic_moment

shielded helion magnetic moments

shielded_proton_magnetic_moment

shielded proton magnetic moments

statampere_square_centimeter

statampere square centimeters

triton_magnetic_moment

triton magnetic moments

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of magnetic moment, IL² (base unit ampere square meter A · m²).

MagneticMoment

Magnetic moment (base unit ampere square meter A · m²).