Module uom::si::mass_concentration

Mass concentration (base unit kilogram per cubic meter, kg · m⁻³).

Structs

attogram_per_cubic_meter

attograms per cubic meter

centigram_per_cubic_meter

centigrams per cubic meter

decagram_per_cubic_meter

decagrams per cubic meter

decigram_per_cubic_meter

decigrams per cubic meter

exagram_per_cubic_meter

exagrams per cubic meter

femtogram_per_cubic_decimeter

femtograms per cubic decimeter

femtogram_per_cubic_meter

femtograms per cubic meter

femtogram_per_deciliter

femtograms per deciliter

femtogram_per_liter

femtograms per liter

femtogram_per_milliliter

femtograms per milliliter

gigagram_per_cubic_meter

gigagrams per cubic meter

gram_per_cubic_decimeter

grams per cubic decimeter

gram_per_cubic_meter

grams per cubic meter

gram_per_deciliter

grams per deciliter

gram_per_liter

grams per liter

gram_per_milliliter

grams per milliliter

hectogram_per_cubic_meter

hectograms per cubic meter

kilogram_per_cubic_decimeter

kilograms per cubic decimeter

kilogram_per_cubic_meter

kilograms per cubic meter

kilogram_per_deciliter

kilograms per deciliter

kilogram_per_liter

kilograms per liter

kilogram_per_milliliter

kilograms per milliliter

megagram_per_cubic_meter

megagrams per cubic meter

microgram_per_cubic_decimeter

micrograms per cubic decimeter

microgram_per_cubic_meter

micrograms per cubic meter

microgram_per_deciliter

micrograms per deciliter

microgram_per_liter

micrograms per liter

microgram_per_milliliter

micrograms per milliliter

milligram_per_cubic_decimeter

milligrams per cubic decimeter

milligram_per_cubic_meter

milligrams per cubic meter

milligram_per_deciliter

milligrams per deciliter

milligram_per_liter

milligrams per liter

milligram_per_milliliter

milligrams per milliliter

nanogram_per_cubic_decimeter

nanograms per cubic decimeter

nanogram_per_cubic_meter

nanograms per cubic meter

nanogram_per_deciliter

nanograms per deciliter

nanogram_per_liter

nanograms per liter

nanogram_per_milliliter

nanograms per milliliter

petagram_per_cubic_meter

petagrams per cubic meter

picogram_per_cubic_decimeter

picograms per cubic decimeter

picogram_per_cubic_meter

picograms per cubic meter

picogram_per_deciliter

picograms per deciliter

picogram_per_liter

picograms per liter

picogram_per_milliliter

picograms per milliliter

teragram_per_cubic_meter

teragrams per cubic meter

yoctogram_per_cubic_meter

yoctograms per cubic meter

yottagram_per_cubic_meter

yottagrams per cubic meter

zeptogram_per_cubic_meter

zeptograms per cubic meter

zettagram_per_cubic_meter

zettagrams per cubic meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of mass concentration, L⁻³M (base unit kilogram per cubic meter, kg · m⁻³).

MassConcentration

Mass concentration (base unit kilogram per cubic meter, kg · m⁻³).