Module uom::si::mass_rate

Mass rate (base unit kilogram per second, kg · s⁻¹).

Structs

attogram_per_second

attograms per second

carat_per_second

carats per second

centigram_per_second

centigrams per second

decagram_per_second

decagrams per second

decigram_per_second

decigrams per second

exagram_per_second

exagrams per second

femtogram_per_second

femtograms per second

gigagram_per_second

gigagrams per second

grain_per_second

grains per second

gram_per_day

grams per day

gram_per_hour

grams per hour

gram_per_minute

grams per minute

gram_per_second

grams per second

hectogram_per_second

hectograms per second

hundredweight_long_per_second

hundredweight (long) per second

hundredweight_short_per_second

hundredweight (short) per second

kilogram_per_day

kilograms per day

kilogram_per_hour

kilograms per hour

kilogram_per_minute

kilograms per minute

kilogram_per_second

Derived unit of mass rate.

megagram_per_second

megagrams per second

microgram_per_second

micrograms per second

milligram_per_second

milligrams per second

nanogram_per_second

nanograms per second

ounce_per_second

ounces per second

ounce_troy_per_second

troy ounces per second

pennyweight_per_second

pennyweight per second

petagram_per_second

petagrams per second

picogram_per_second

picograms per second

pound_per_day

pounds per day

pound_per_hour

pounds per hour

pound_per_minute

pounds per minute

pound_per_second

pounds per second

pound_troy_per_second

troy pounds per second

slug_per_second

slugs per second

teragram_per_second

teragrams per second

ton_assay_per_second

assay tons per second

ton_long_per_second

long tons per second

ton_per_day

Ton per day, metric.

ton_per_hour

Ton per hour, metric.

ton_per_minute

Ton per minute, metric.

ton_per_second

Ton per second, metric.

ton_short_per_hour

short tons per hour

ton_short_per_second

short tons per second

yoctogram_per_second

yoctograms per second

yottagram_per_second

yottagrams per second

zeptogram_per_second

zeptograms per second

zettagram_per_second

zettagrams per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of mass rate, MT⁻¹ (base unit kilogram per second, kg · s⁻¹).

MassRate

Mass rate (base unit kilogram per second, kg · s⁻¹).