Module uom::si::molar_concentration

Molar concentration (base unit mole per cubic meter, mol · m⁻³).

Structs

attomole_per_cubic_meter

attomoles per cubic meter

centimole_per_cubic_meter

centimoles per cubic meter

decamole_per_cubic_meter

decamoles per cubic meter

decimole_per_cubic_meter

decimoles per cubic meter

examole_per_cubic_meter

examoles per cubic meter

femtomole_per_cubic_decimeter

femtomoles per cubic decimeter

femtomole_per_cubic_meter

femtomoles per cubic meter

femtomole_per_deciliter

femtomoles per deciliter

femtomole_per_liter

femtomoles per liter

femtomole_per_milliliter

femtomoles per milliliter

gigamole_per_cubic_meter

gigamoles per cubic meter

hectomole_per_cubic_meter

hectomoles per cubic meter

kilomole_per_cubic_decimeter

kilomoles per cubic decimeter

kilomole_per_cubic_meter

kilomoles per cubic meter

kilomole_per_deciliter

kilomoles per deciliter

kilomole_per_liter

kilomoles per liter

kilomole_per_milliliter

kilomoles per milliliter

megamole_per_cubic_meter

megamoles per cubic meter

micromole_per_cubic_decimeter

micromoles per cubic decimeter

micromole_per_cubic_meter

micromoles per cubic meter

micromole_per_deciliter

micromoles per deciliter

micromole_per_liter

micromoles per liter

micromole_per_milliliter

micromoles per milliliter

millimole_per_cubic_decimeter

millimoles per cubic decimeter

millimole_per_cubic_meter

millimoles per cubic meter

millimole_per_deciliter

millimoles per deciliter

millimole_per_liter

millimoles per liter

millimole_per_milliliter

millimoles per milliliter

mole_per_cubic_decimeter

moles per cubic decimeter

mole_per_cubic_meter

moles per cubic meter

mole_per_deciliter

moles per deciliter

mole_per_liter

moles per liter

mole_per_milliliter

moles per milliliter

nanomole_per_cubic_decimeter

nanomoles per cubic decimeter

nanomole_per_cubic_meter

nanomoles per cubic meter

nanomole_per_deciliter

nanomoles per deciliter

nanomole_per_liter

nanomoles per liter

nanomole_per_milliliter

nanomoles per milliliter

particle_per_cubic_decimeter

particles per cubic decimeter

particle_per_cubic_meter

particles per cubic meter

particle_per_deciliter

particles per deciliter

particle_per_liter

particles per liter

particle_per_milliliter

particles per milliliter

petamole_per_cubic_meter

petamoles per cubic meter

picomole_per_cubic_decimeter

picomoles per cubic decimeter

picomole_per_cubic_meter

picomoles per cubic meter

picomole_per_deciliter

picomoles per deciliter

picomole_per_liter

picomoles per liter

picomole_per_milliliter

picomoles per milliliter

teramole_per_cubic_meter

teramoles per cubic meter

yoctomole_per_cubic_meter

yoctomoles per cubic meter

yottamole_per_cubic_meter

yottamoles per cubic meter

zeptomole_per_cubic_meter

zeptomoles per cubic meter

zettamole_per_cubic_meter

zettamoles per cubic meter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of molar concentration, L⁻³N (base unit mole per cubic meter, mol · m⁻³).

MolarConcentration

Molar concentration (base unit mole per cubic meter, mol · m⁻³).