Module uom::si::molar_heat_capacity

Molar heat capacity (base unit joule per kelvin mole, m² · kg · s⁻² · K⁻¹ · mol⁻¹).

Structs

attojoule_per_kelvin_mole

attojoules per kelvin mole

btu_39_per_kelvin_mole

British thermal units (39 °F) per kelvin mole

btu_59_per_kelvin_mole

British thermal units (59 °F) per kelvin mole

btu_60_per_kelvin_mole

British thermal units (60 °F) per kelvin mole

btu_it_per_kelvin_mole

British thermal units (IT) per kelvin mole

btu_per_kelvin_mole

British thermal units per kelvin mole

calorie_15_per_kelvin_mole

calories (15 °C) per kelvin mole

calorie_20_per_kelvin_mole

calories (20 °C) per kelvin mole

calorie_it_nutrition_per_kelvin_mole

Calories (IT) per kelvin mole

calorie_it_per_kelvin_mole

calories (IT) per kelvin mole

calorie_nutrition_per_kelvin_mole

Calories per kelvin mole

calorie_per_kelvin_mole

calories per kelvin mole

centijoule_per_kelvin_mole

centijoules per kelvin mole

decajoule_per_kelvin_mole

decajoules per kelvin mole

decijoule_per_kelvin_mole

decijoules per kelvin mole

electronvolt_per_kelvin_particle

electronvolts per kelvin particle

exajoule_per_kelvin_mole

exajoules per kelvin mole

femtojoule_per_kelvin_mole

femtojoules per kelvin mole

gigajoule_per_kelvin_mole

gigajoules per kelvin mole

hectojoule_per_kelvin_mole

hectojoules per kelvin mole

joule_per_kelvin_mole

Derived unit of molar heat capacity.

joule_per_kelvin_particle

joules per kelvin particle

kilocalorie_it_per_kelvin_mole

kilocalories (IT) per kelvin mole

kilocalorie_per_kelvin_mole

kilocalories per kelvin mole

kilojoule_per_kelvin_mole

kilojoules per kelvin mole

megajoule_per_kelvin_mole

megajoules per kelvin mole

microjoule_per_kelvin_mole

microjoules per kelvin mole

millijoule_per_kelvin_mole

millijoules per kelvin mole

molar_gas_constant

molar gas constants

nanojoule_per_kelvin_mole

nanojoules per kelvin mole

petajoule_per_kelvin_mole

petajoules per kelvin mole

picojoule_per_kelvin_mole

picojoules per kelvin mole

terajoule_per_kelvin_mole

terajoules per kelvin mole

yoctojoule_per_kelvin_mole

yoctojoules per kelvin mole

yottajoule_per_kelvin_mole

yottajoules per kelvin mole

zeptojoule_per_kelvin_mole

zeptojoules per kelvin mole

zettajoule_per_kelvin_mole

zettajoules per kelvin mole

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of molar heat capacity, L²MT⁻²Th⁻¹N⁻¹ (base unit joule per kelvin mole, m² · kg · s⁻² · K⁻¹ · mol⁻¹).

MolarHeatCapacity

Molar heat capacity (base unit joule per kelvin mole, m² · kg · s⁻² · K⁻¹ · mol⁻¹).