Module uom::si::molar_mass

Molar mass (base unit kilogram per mole, kg · mol⁻¹).

Structs

attogram_per_mole
Measurement unit.

centigram_per_mole
Measurement unit.

decagram_per_mole
Measurement unit.

decigram_per_mole
Measurement unit.

exagram_per_mole
Measurement unit.

femtogram_per_mole
Measurement unit.

gigagram_per_mole
Measurement unit.

gram_per_mole
Measurement unit.

hectogram_per_mole
Measurement unit.

kilogram_per_mole
The derived unit of molar mass.

megagram_per_mole
Measurement unit.

microgram_per_mole
Measurement unit.

milligram_per_mole
Measurement unit.

nanogram_per_mole
Measurement unit.

petagram_per_mole
Measurement unit.

picogram_per_mole
Measurement unit.

teragram_per_mole
Measurement unit.

yoctogram_per_mole
Measurement unit.

yottagram_per_mole
Measurement unit.

zeptogram_per_mole
Measurement unit.

zettagram_per_mole
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of molar mass, MN⁻¹ (base unit kilogram per mole, kg · mol⁻¹).

MolarMass
Molar mass (base unit kilogram per mole, kg · mol⁻¹).