Module uom::si::momentum

Momentum (base unit kilogram meter per second, kg · m · s⁻¹).

Structs

attogram_meter_per_second

attogram meters per second

centigram_meter_per_second

centigram meters per second

decagram_meter_per_second

decagram meters per second

decigram_meter_per_second

decigram meters per second

exagram_meter_per_second

exagram meters per second

femtogram_meter_per_second

femtogram meters per second

gigagram_meter_per_second

gigagram meters per second

gram_meter_per_second

gram meters per second

hectogram_meter_per_second

hectogram meters per second

kilogram_attometer_per_second

kilogram attometers per second

kilogram_centimeter_per_second

kilogram centimeters per second

kilogram_decameter_per_second

kilogram decameters per second

kilogram_decimeter_per_second

kilogram decimeters per second

kilogram_exameter_per_second

kilogram exameters per second

kilogram_femtometer_per_second

kilogram femtometers per second

kilogram_gigameter_per_second

kilogram gigameters per second

kilogram_hectometer_per_second

kilogram hectometers per second

kilogram_kilometer_per_second

kilogram kilometers per second

kilogram_megameter_per_second

kilogram megameters per second

kilogram_meter_per_day

kilogram meters per day

kilogram_meter_per_hour

kilogram meters per hour

kilogram_meter_per_minute

kilogram meters per minute

kilogram_meter_per_second

Derived unit of momentum.

kilogram_micrometer_per_second

kilogram micrometers per second

kilogram_millimeter_per_second

kilogram millimeters per second

kilogram_nanometer_per_second

kilogram nanometers per second

kilogram_petameter_per_second

kilogram petameters per second

kilogram_picometer_per_second

kilogram picometers per second

kilogram_terameter_per_second

kilogram terameters per second

kilogram_yoctometer_per_second

kilogram yoctometers per second

kilogram_yottameter_per_second

kilogram yottameters per second

kilogram_zeptometer_per_second

kilogram zeptometers per second

kilogram_zettameter_per_second

kilogram zettameters per second

megagram_meter_per_second

megagram meters per second

microgram_meter_per_second

microgram meters per second

milligram_meter_per_second

milligram meters per second

nanogram_meter_per_second

nanogram meters per second

petagram_meter_per_second

petagram meters per second

picogram_meter_per_second

picogram meters per second

pound_foot_per_second

pound feet per second

pound_inch_per_second

pound inches per second

slug_foot_per_second

slug feet per second

slug_inch_per_second

slug inches per second

teragram_meter_per_second

teragram meters per second

ton_meter_per_second

ton meters per second

yoctogram_meter_per_second

yoctogram meters per second

yottagram_meter_per_second

yottagram meters per second

zeptogram_meter_per_second

zeptogram meters per second

zettagram_meter_per_second

zettagram meters per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of momentum, LMT⁻¹ (base unit kilogram meter per second, kg · m · s⁻¹).

Momentum

Momentum (base unit kilogram meter per second, kg · m · s⁻¹).