Module uom::si::power

Power (base unit watt, m² · kg · s⁻³).

Structs

attowatt

attowatts

centiwatt

centiwatts

decawatt

decawatts

deciwatt

deciwatts

erg_per_second

ergs per second

exawatt

exawatts

femtowatt

femtowatts

foot_pound_per_hour

foot pounds-force per hour

foot_pound_per_minute

foot pounds-force per minute

foot_pound_per_second

foot pounds-force per second

gigawatt

gigawatts

hectowatt

hectowatts

horsepower

horsepower

horsepower_boiler

horsepower (boiler)

horsepower_electric

horsepower (electric)

horsepower_imperial

horsepower (Imperial)

horsepower_metric

metric horsepower

hydraulic_horsepower

hydraulic horsepower

kilowatt

kilowatts

megawatt

megawatts

microwatt

microwatts

milliwatt

milliwatts

nanowatt

nanowatts

petawatt

petawatts

picowatt

picowatts

terawatt

terawatts

watt

Derived unit of power.

yoctowatt

yoctowatts

yottawatt

yottawatts

zeptowatt

zeptowatts

zettawatt

zettawatts

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of power, L²MT⁻³ (base unit watt, m² · kg · s⁻³).

Power

Power (base unit watt, m² · kg · s⁻³).