Module uom::si::quantities

Quantity type aliases using the default base units and parameterized on the underlying storage type.

Type Aliases

Absement

Quantity type alias using the default base units parameterized on the underlying storage type.

Acceleration

Quantity type alias using the default base units parameterized on the underlying storage type.

Action

Quantity type alias using the default base units parameterized on the underlying storage type.

AmountOfSubstance

Quantity type alias using the default base units parameterized on the underlying storage type.

Angle

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularAbsement

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularAcceleration

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularJerk

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularVelocity

Quantity type alias using the default base units parameterized on the underlying storage type.

Area

Quantity type alias using the default base units parameterized on the underlying storage type.

ArealDensityOfStates

Quantity type alias using the default base units parameterized on the underlying storage type.

ArealHeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

ArealMassDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ArealNumberDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ArealNumberRate

Quantity type alias using the default base units parameterized on the underlying storage type.

AvailableEnergy

Quantity type alias using the default base units parameterized on the underlying storage type.

Capacitance

Quantity type alias using the default base units parameterized on the underlying storage type.

CatalyticActivity

Quantity type alias using the default base units parameterized on the underlying storage type.

CatalyticActivityConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

Curvature

Quantity type alias using the default base units parameterized on the underlying storage type.

DiffusionCoefficient

Quantity type alias using the default base units parameterized on the underlying storage type.

DynamicViscosity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCharge

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricChargeArealDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricChargeLinearDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricChargeVolumetricDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCurrent

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCurrentDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricDipoleMoment

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricDisplacementField

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricField

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricFlux

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricPermittivity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricPotential

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricQuadrupoleMoment

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalConductance

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalConductivity

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalMobility

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalResistance

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalResistivity

Quantity type alias using the default base units parameterized on the underlying storage type.

Energy

Quantity type alias using the default base units parameterized on the underlying storage type.

Force

Quantity type alias using the default base units parameterized on the underlying storage type.

Frequency

Quantity type alias using the default base units parameterized on the underlying storage type.

FrequencyDrift

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatFluxDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatTransfer

Quantity type alias using the default base units parameterized on the underlying storage type.

Inductance

Quantity type alias using the default base units parameterized on the underlying storage type.

Information

Quantity type alias using the default base units parameterized on the underlying storage type.

InformationRate

Quantity type alias using the default base units parameterized on the underlying storage type.

InverseVelocity

Quantity type alias using the default base units parameterized on the underlying storage type.

Jerk

Quantity type alias using the default base units parameterized on the underlying storage type.

Length

Quantity type alias using the default base units parameterized on the underlying storage type.

LinearDensityOfStates

Quantity type alias using the default base units parameterized on the underlying storage type.

LinearMassDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

LinearNumberDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

LinearNumberRate

Quantity type alias using the default base units parameterized on the underlying storage type.

LinearPowerDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

Luminance

Quantity type alias using the default base units parameterized on the underlying storage type.

LuminousIntensity

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFieldStrength

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFlux

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFluxDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticMoment

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticPermeability

Quantity type alias using the default base units parameterized on the underlying storage type.

Mass

Quantity type alias using the default base units parameterized on the underlying storage type.

MassConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

MassDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

MassFlux

Quantity type alias using the default base units parameterized on the underlying storage type.

MassPerEnergy

Quantity type alias using the default base units parameterized on the underlying storage type.

MassRate

Quantity type alias using the default base units parameterized on the underlying storage type.

Molality

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarEnergy

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarFlux

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarHeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarMass

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarRadioactivity

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarVolume

Quantity type alias using the default base units parameterized on the underlying storage type.

MomentOfInertia

Quantity type alias using the default base units parameterized on the underlying storage type.

Momentum

Quantity type alias using the default base units parameterized on the underlying storage type.

Power

Quantity type alias using the default base units parameterized on the underlying storage type.

PowerRate

Quantity type alias using the default base units parameterized on the underlying storage type.

Pressure

Quantity type alias using the default base units parameterized on the underlying storage type.

RadiantExposure

Quantity type alias using the default base units parameterized on the underlying storage type.

Radioactivity

Quantity type alias using the default base units parameterized on the underlying storage type.

Ratio

Quantity type alias using the default base units parameterized on the underlying storage type.

ReciprocalLength

Quantity type alias using the default base units parameterized on the underlying storage type.

SolidAngle

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificArea

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificHeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificPower

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificRadioactivity

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificVolume

Quantity type alias using the default base units parameterized on the underlying storage type.

SurfaceElectricCurrentDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

TemperatureCoefficient

Quantity type alias using the default base units parameterized on the underlying storage type.

TemperatureGradient

Quantity type alias using the default base units parameterized on the underlying storage type.

TemperatureInterval

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermalConductance

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermalConductivity

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermalResistance

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermodynamicTemperature

Quantity type alias using the default base units parameterized on the underlying storage type.

Time

Quantity type alias using the default base units parameterized on the underlying storage type.

Torque

Quantity type alias using the default base units parameterized on the underlying storage type.

Velocity

Quantity type alias using the default base units parameterized on the underlying storage type.

Volume

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumeRate

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumetricDensityOfStates

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumetricHeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumetricNumberDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumetricNumberRate

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumetricPowerDensity

Quantity type alias using the default base units parameterized on the underlying storage type.