Module uom::si::ratio

Ratio (dimensionless quantity).

Structs

basis_point
Measurement unit.

part_per_billion
Measurement unit.

part_per_hundred
Measurement unit.

part_per_million
Measurement unit.

part_per_quadrillion
Measurement unit.

part_per_ten_thousand
Measurement unit.

part_per_thousand
Measurement unit.

part_per_trillion
Measurement unit.

per_mille
Measurement unit.

percent
Measurement unit.

ratio
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of ratio, 1 (dimensionless).

Ratio
Ratio (dimensionless quantity).