Module uom::si::specific_power

Specific power (base unit watt per kilogram, m² · s⁻³).

Structs

attowatt_per_kilogram

attowatts per kilogram

centiwatt_per_kilogram

centiwatts per kilogram

decawatt_per_kilogram

decawatts per kilogram

deciwatt_per_kilogram

deciwatts per kilogram

erg_per_second_per_kilogram

ergs per second per kilogram

exawatt_per_kilogram

exawatts per kilogram

femtowatt_per_kilogram

femtowatts per kilogram

foot_pound_per_hour_per_pound

foot pounds-force per hour per pound

foot_pound_per_minute_per_pound

foot pounds-force per minute per pound

foot_pound_per_second_per_pound

foot pounds-force per second per pound

gigawatt_per_kilogram

gigawatts per kilogram

hectowatt_per_kilogram

hectowatts per kilogram

horsepower_boiler_per_pound

horsepower (boiler) per pound

horsepower_electric_per_pound

horsepower (electric) per pound

horsepower_imperial_per_pound

horsepower (Imperial) per pound

horsepower_metric_per_pound

metric horsepower per pound

horsepower_per_pound

horsepower per pound

hydraulic_horsepower_per_pound

hydraulic horsepower per pound

kilowatt_per_kilogram

kilowatts per kilogram

megawatt_per_kilogram

megawatts per kilogram

microwatt_per_kilogram

microwatts per kilogram

milliwatt_per_kilogram

milliwatts per kilogram

nanowatt_per_kilogram

nanowatts per kilogram

petawatt_per_kilogram

petawatts per kilogram

picowatt_per_kilogram

picowatts per kilogram

square_meter_per_cubic_second

Less commonly used, but more accurate (reduced) derived unit of specific power.

terawatt_per_kilogram

terawatts per kilogram

watt_per_kilogram

Derived unit of specific power.

yoctowatt_per_kilogram

yoctowatts per kilogram

yottawatt_per_kilogram

yottawatts per kilogram

zeptowatt_per_kilogram

zeptowatts per kilogram

zettawatt_per_kilogram

zettawatts per kilogram

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of specific power, L²T⁻³ (base unit watt per kilogram, m² · s⁻³).

SpecificPower

Specific power (base unit watt per kilogram, m² · s⁻³).