Struct uom::si::Quantity

#[repr(transparent)]
pub struct Quantity<D: ?Sized, U: ?Sized, V> where
    D: Dimension,
    U: Units<V>,
    V: Num + Conversion<V>,  {
    pub dimension: PhantomData<D>,
    pub units: PhantomData<U>,
    pub value: V,
}

Property of a phenomenon, body or substance, where the property has a magnitude that can be expressed as a number and a reference.

The preferred method of creating a Quantity instance is to use the new constructor which is generic over the input unit and accepts the input value as it's only parameter.

// Create a length of 1 meter.
let l = Length::new::<meter>(1.0);

Quantity fields are public to allow for the creation of const values and instances of non-named Quantitys. This functionality will be deprecated and subsequently removed once the const fn feature is stabilized.

// Create a `const` length of 1 meter.
const L: Length = Length { dimension: PhantomData, units: PhantomData, value: 1.0, };
// Create a `const` area of 1 square meter explicitly without using the `Area` alias.
const A: Quantity<ISQ<P2, Z0, Z0, Z0, Z0, Z0, Z0>, SI<f32>, f32> =
    Quantity { dimension: PhantomData, units: PhantomData, value: 1.0, };

Using units for the wrong quantity will cause a compile error:

// error[E0277]: the trait bound `second: length::Unit` is not satisfied
let l = Length::new::<second>(1.0);

Mixing quantities will also cause a compile error:

// error[E0308]: mismatched types
let r = Length::new::<meter>(1.0) + Time::new::<second>(1.0);

// error[E0308]: mismatched types
let v: Velocity = Length::new::<meter>(1.0) * Time::new::<second>(1.0);

http://jcgm.bipm.org/vim/en/1.1.html

Generic Parameters

D: Quantity dimension. See Dimension.

U: Quantity base units. See Units.

V: Quantity value underlying storage type.

Fields
dimension: PhantomData<D>

Quantity dimension. See Dimension.

units: PhantomData<U>

Quantity base units. See Units.

value: V

Quantity value stored in the base units for the quantity.

Methods

impl<D: ?Sized, U: ?Sized, V> Quantity<D, U, V> where
D: Dimension,
U: Units<V>,
V: Num + Conversion<V>,

pub fn is_nan(self) -> bool where
V: Float,

Returns true if this value is NAN and false otherwise.

pub fn is_infinite(self) -> bool where
V: Float,

Returns true if this value is positive infinity or negative infinity and false otherwise.

pub fn is_finite(self) -> bool where
V: Float,

Returns true if this number is neither infinite nor NAN.

pub fn is_normal(self) -> bool where
V: Float,

Returns true if the number is neither zero, infinite, subnormal, or NAN.

pub fn classify(self) -> FpCategory where
V: Float,

Returns the floating point category of the number. If only one property is going to be tested, it is generally faster to use the specific predicate instead.

pub fn cbrt(
self
) -> Quantity<ISQ<PartialQuot<D::L, P3>, PartialQuot<D::M, P3>, PartialQuot<D::T, P3>, PartialQuot<D::I, P3>, PartialQuot<D::Th, P3>, PartialQuot<D::N, P3>, PartialQuot<D::J, P3>>, U, V> where
D::L: PartialDiv<P3>,
D::M: PartialDiv<P3>,
D::T: PartialDiv<P3>,
D::I: PartialDiv<P3>,
D::Th: PartialDiv<P3>,
D::N: PartialDiv<P3>,
D::J: PartialDiv<P3>,
D::Kind: Div,
V: Float,

Takes the cubic root of a number.

let l: Length = Volume::new::<cubic_meter>(8.0).cbrt();

The input type must have dimensions divisible by three:

// error[E0271]: type mismatch resolving ...
let r = Area::new::<square_meter>(8.0).cbrt();

pub fn abs(self) -> Self where
V: Signed,

Computes the absolute value of self. Returns NAN if the quantity is NAN.

pub fn signum(self) -> Self where
V: Signed,

Returns a quantity that represents the sign of self.

1.0 of the base unit if the number is positive, +0.0, or INFINITY.

-1.0 of the base unit if the number is negative, -0.0, or NEG_INFINITY.

NAN if the number is NAN.

pub fn is_sign_positive(self) -> bool where
V: Float,

Returns true if self's sign bit is positive, including +0.0 and INFINITY.

pub fn is_sign_negative(self) -> bool where
V: Float,

Returns true if self's sign is negative, including -0.0 and NEG_INFINITY.

pub fn mul_add<Da: ?Sized, Ua: ?Sized, Ub: ?Sized>(
self,
a: Quantity<Da, Ua, V>,
b: Quantity<ISQ<Sum<D::L, Da::L>, Sum<D::M, Da::M>, Sum<D::T, Da::T>, Sum<D::I, Da::I>, Sum<D::Th, Da::Th>, Sum<D::N, Da::N>, Sum<D::J, Da::J>>, Ub, V>
) -> Quantity<ISQ<Sum<D::L, Da::L>, Sum<D::M, Da::M>, Sum<D::T, Da::T>, Sum<D::I, Da::I>, Sum<D::Th, Da::Th>, Sum<D::N, Da::N>, Sum<D::J, Da::J>>, U, V> where
D::L: Add<Da::L>,
D::M: Add<Da::M>,
D::T: Add<Da::T>,
D::I: Add<Da::I>,
D::Th: Add<Da::Th>,
D::N: Add<Da::N>,
D::J: Add<Da::J>,
D::Kind: Mul,
V: Float,
Da: Dimension,
Da::Kind: Mul,
Ua: Units<V>,
Ub: Units<V>,

Fused multiply-add. Computes (self * a) + b with only one rounding error. This produces a more accurate result with better performance than a separate multiplication operation followed by an add.

Generic Parameters

Da: Dimension for parameter a.

Ua: Base units for parameter a.

Ub: Base units for parameter b.

pub fn recip(
self
) -> Quantity<ISQ<Negate<D::L>, Negate<D::M>, Negate<D::T>, Negate<D::I>, Negate<D::Th>, Negate<D::N>, Negate<D::J>>, U, V> where
D::L: Neg,
D::M: Neg,
D::T: Neg,
D::I: Neg,
D::Th: Neg,
D::N: Neg,
D::J: Neg,
D::Kind: Div,
V: Float,

Takes the reciprocal (inverse) of a number, 1/x.

let f: Frequency = Time::new::<second>(1.0).recip();

pub fn powi<E>(
self,
e: E
) -> Quantity<ISQ<Prod<D::L, E>, Prod<D::M, E>, Prod<D::T, E>, Prod<D::I, E>, Prod<D::Th, E>, Prod<D::N, E>, Prod<D::J, E>>, U, V> where
D::L: Mul<E>,
D::M: Mul<E>,
D::T: Mul<E>,
D::I: Mul<E>,
D::Th: Mul<E>,
D::N: Mul<E>,
D::J: Mul<E>,
D::Kind: Mul,
E: Integer,
V: Pow<E, Output = V> + Conversion<V>,

Raises a quantity to an integer power.

let a: Area = Length::new::<meter>(3.0).powi(::uom::typenum::P2::new());

Generic Parameters

E: typenum::Integer power.

pub fn sqrt(
self
) -> Quantity<ISQ<PartialQuot<D::L, P2>, PartialQuot<D::M, P2>, PartialQuot<D::T, P2>, PartialQuot<D::I, P2>, PartialQuot<D::Th, P2>, PartialQuot<D::N, P2>, PartialQuot<D::J, P2>>, U, V> where
D::L: PartialDiv<P2>,
D::M: PartialDiv<P2>,
D::T: PartialDiv<P2>,
D::I: PartialDiv<P2>,
D::Th: PartialDiv<P2>,
D::N: PartialDiv<P2>,
D::J: PartialDiv<P2>,
D::Kind: Div,
V: Float,

Takes the square root of a number. Returns NAN if self is a negative number.

let l: Length = Area::new::<square_meter>(4.0).sqrt();

The input type must have dimensions divisible by two:

// error[E0271]: type mismatch resolving ...
let r = Length::new::<meter>(4.0).sqrt();

pub fn max(self, other: Self) -> Self where
V: Float,

Returns the maximum of the two quantities.

pub fn min(self, other: Self) -> Self where
V: Float,

Returns the minimum of the two quantities.