Module uom::si::temperature_interval

Temperature interval (base unit kelvin, K).

Temperature interval has the same dimensions as thermodynamic temperature but is not directly comparable. See thermodynamic temperature for a full explanation.

Structs

attokelvin

attokelvins

centikelvin

centikelvins

decakelvin

decakelvins

decikelvin

decikelvins

degree_celsius

degrees Celsius

degree_fahrenheit

degrees Fahrenheit

degree_rankine

degrees Rankine

exakelvin

exakelvins

femtokelvin

femtokelvins

gigakelvin

gigakelvins

hectokelvin

hectokelvins

kelvin

The kelvin is the SI unit of thermodynamic temperature. It is defined by taking the fixed numerical value of the Boltzmann constant k to be 1.380 649 × 10⁻²³ when expressed in the unit J K⁻¹, which is equal to kg m² s⁻² K⁻¹, where the kilogram, meter, and second are defined in terms of h, c, and ∆νCs.

kilokelvin

kilokelvins

megakelvin

megakelvins

microkelvin

microkelvins

millikelvin

millikelvins

nanokelvin

nanokelvins

petakelvin

petakelvins

picokelvin

picokelvins

terakelvin

terakelvins

yoctokelvin

yoctokelvins

yottakelvin

yottakelvins

zeptokelvin

zeptokelvins

zettakelvin

zettakelvins

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of temperature interval, Th (base unit kelvin, K).

TemperatureInterval

Temperature interval (base unit kelvin, K).