Module uom::si::thermal_conductance

Thermal conductance (base unit watt per kelvin, kg · m² · s⁻³ · K⁻¹).

Thermal conductance has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attogram_meter_squared_per_second_cubed_kelvin

attograms meter squared per second cubed kelvin

attowatt_per_kelvin

attowatts per kelvin

centigram_meter_squared_per_second_cubed_kelvin

centigrams meter squared per second cubed kelvin

centiwatt_per_kelvin

centiwatts per kelvin

decagram_meter_squared_per_second_cubed_kelvin

decagrams meter squared per second cubed kelvin

decawatt_per_kelvin

decawatts per kelvin

decigram_meter_squared_per_second_cubed_kelvin

decigrams meter squared per second cubed kelvin

deciwatt_per_kelvin

deciwatts per kelvin

exagram_meter_squared_per_second_cubed_kelvin

exagrams meter squared per second cubed kelvin

exawatt_per_kelvin

exawatts per kelvin

femtogram_meter_squared_per_second_cubed_kelvin

femtograms meter squared per second cubed kelvin

femtowatt_per_kelvin

femtowatts per kelvin

gigagram_meter_squared_per_second_cubed_kelvin

gigagrams meter squared per second cubed kelvin

gigawatt_per_kelvin

gigawatts per kelvin

gram_meter_squared_per_second_cubed_kelvin

grams meter squared per second cubed kelvin

hectogram_meter_squared_per_second_cubed_kelvin

hectograms meter squared per second cubed kelvin

hectowatt_per_kelvin

hectowatts per kelvin

kilogram_meter_squared_per_second_cubed_degree_celsius

kilograms meter squared per second cubed degree Celsius

kilogram_meter_squared_per_second_cubed_kelvin

Derived unit of thermal conductance in base units. Equivalent to W/K.

kilowatt_per_degree_celsius

kilowatts per degree Celsius

kilowatt_per_kelvin

kilowatts per kelvin

megagram_meter_squared_per_second_cubed_kelvin

megagrams meter squared per second cubed kelvin

megawatt_per_kelvin

megawatts per kelvin

microgram_meter_squared_per_second_cubed_kelvin

micrograms meter squared per second cubed kelvin

microwatt_per_kelvin

microwatts per kelvin

milligram_meter_squared_per_second_cubed_kelvin

milligrams meter squared per second cubed kelvin

milliwatt_per_degree_celsius

milliwatts per degree Celsius

milliwatt_per_kelvin

milliwatts per kelvin

nanogram_meter_squared_per_second_cubed_kelvin

nanograms meter squared per second cubed kelvin

nanowatt_per_kelvin

nanowatts per kelvin

petagram_meter_squared_per_second_cubed_kelvin

petagrams meter squared per second cubed kelvin

petawatt_per_kelvin

petawatts per kelvin

picogram_meter_squared_per_second_cubed_kelvin

picograms meter squared per second cubed kelvin

picowatt_per_kelvin

picowatts per kelvin

teragram_meter_squared_per_second_cubed_kelvin

teragrams meter squared per second cubed kelvin

terawatt_per_kelvin

terawatts per kelvin

watt_per_kelvin

Derived unit of thermal conductance in derived units. Equivalent to kg · m²/(s³ · K).

watt_per_meter_degree_celsius

Derived unit of thermal conductance in derived units. Equivalent to kg · m²/(s³ · K).

yoctogram_meter_squared_per_second_cubed_kelvin

yoctograms meter squared per second cubed kelvin

yoctowatt_per_kelvin

yoctowatts per kelvin

yottagram_meter_squared_per_second_cubed_kelvin

yottagrams meter squared per second cubed kelvin

yottawatt_per_kelvin

yottawatts per kelvin

zeptogram_meter_squared_per_second_cubed_kelvin

zeptograms meter squared per second cubed kelvin

zeptowatt_per_kelvin

zeptowatts per kelvin

zettagram_meter_squared_per_second_cubed_kelvin

zettagrams meter squared per second cubed kelvin

zettawatt_per_kelvin

zettawatts per kelvin

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of thermal conductance, LM²T⁻³Th⁻¹ (base unit watt per kelvin, kg · m² · s⁻³ · K⁻¹).

ThermalConductance

Thermal conductance (base unit watt per kelvin, kg · m² · s⁻³ · K⁻¹).