Module uom::si::thermal_conductivity

Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

Thermal conductivity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attogram_meter_per_second_cubed_kelvin

attograms meter per second cubed kelvin

attowatt_per_meter_kelvin

attowatts per meter kelvin

centigram_meter_per_second_cubed_kelvin

centigrams meter per second cubed kelvin

centiwatt_per_meter_kelvin

centiwatts per meter kelvin

decagram_meter_per_second_cubed_kelvin

decagrams meter per second cubed kelvin

decawatt_per_meter_kelvin

decawatts per meter kelvin

decigram_meter_per_second_cubed_kelvin

decigrams meter per second cubed kelvin

deciwatt_per_meter_kelvin

deciwatts per meter kelvin

exagram_meter_per_second_cubed_kelvin

exagrams meter per second cubed kelvin

exawatt_per_meter_kelvin

exawatts per meter kelvin

femtogram_meter_per_second_cubed_kelvin

femtograms meter per second cubed kelvin

femtowatt_per_meter_kelvin

femtowatts per meter kelvin

gigagram_meter_per_second_cubed_kelvin

gigagrams meter per second cubed kelvin

gigawatt_per_meter_kelvin

gigawatts per meter kelvin

gram_meter_per_second_cubed_kelvin

grams meter per second cubed kelvin

hectogram_meter_per_second_cubed_kelvin

hectograms meter per second cubed kelvin

hectowatt_per_meter_kelvin

hectowatts per meter kelvin

kilogram_meter_per_second_cubed_degree_celsius

kilograms meter per second cubed degree celsius

kilogram_meter_per_second_cubed_kelvin

Derived unit of thermal conductivity in base units. Equivalent to W/(m · K).

kilowatt_per_meter_degree_celsius

kilowatts per meter degree celsius

kilowatt_per_meter_kelvin

kilowatts per meter kelvin

megagram_meter_per_second_cubed_kelvin

megagrams meter per second cubed kelvin

megawatt_per_meter_kelvin

megawatts per meter kelvin

microgram_meter_per_second_cubed_kelvin

micrograms meter per second cubed kelvin

microwatt_per_meter_kelvin

microwatts per meter kelvin

milligram_meter_per_second_cubed_kelvin

milligrams meter per second cubed kelvin

milliwatt_per_meter_degree_celsius

milliwatts per meter degree celsius

milliwatt_per_meter_kelvin

milliwatts per meter kelvin

nanogram_meter_per_second_cubed_kelvin

nanograms meter per second cubed kelvin

nanowatt_per_meter_kelvin

nanowatts per meter kelvin

petagram_meter_per_second_cubed_kelvin

petagrams meter per second cubed kelvin

petawatt_per_meter_kelvin

petawatts per meter kelvin

picogram_meter_per_second_cubed_kelvin

picograms meter per second cubed kelvin

picowatt_per_meter_kelvin

picowatts per meter kelvin

teragram_meter_per_second_cubed_kelvin

teragrams meter per second cubed kelvin

terawatt_per_meter_kelvin

terawatts per meter kelvin

watt_per_meter_degree_celsius

Derived unit of thermal conductivity in derived units. Equivalent to kg · m/(s³ · K).

watt_per_meter_kelvin

Derived unit of thermal conductivity in derived units. Equivalent to kg · m/(s³ · K).

yoctogram_meter_per_second_cubed_kelvin

yoctograms meter per second cubed kelvin

yoctowatt_per_meter_kelvin

yoctowatts per meter kelvin

yottagram_meter_per_second_cubed_kelvin

yottagrams meter per second cubed kelvin

yottawatt_per_meter_kelvin

yottawatts per meter kelvin

zeptogram_meter_per_second_cubed_kelvin

zeptograms meter per second cubed kelvin

zeptowatt_per_meter_kelvin

zeptowatts per meter kelvin

zettagram_meter_per_second_cubed_kelvin

zettagrams meter per second cubed kelvin

zettawatt_per_meter_kelvin

zettawatts per meter kelvin

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of thermal conductivity, LMT⁻³Th⁻¹ (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

ThermalConductivity

Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).