Module uom::si::thermodynamic_temperature

Thermodynamic temperature (base unit kelvin, K).

Thermodynamic temperature has the same dimensions as temperature interval but is not directly comparable. Thermodynamic temperature is the absolute measure of temperature and is one of the base quantities in the ISQ. Temperature interval is the measure of relative temperature difference between thermodynamic temperatures.

let tt = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let ti = TemperatureInterval::new::<temperature_interval::kelvin>(1.0);

// error[E0308]: mismatched types
let err = tt == ti;

Additionally, addition and subtraction are not implemented for thermodynamic temperature.

let t1 = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let t2 = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);

// error[E0308]: mismatched types
let err = t1 + t2;

A temperature interval can be added to or subtracted from a thermodynamic temperature.

let tt = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let ti = TemperatureInterval::new::<temperature_interval::kelvin>(1.0);

let result = tt + ti;

Structs

attokelvin

attokelvins

centikelvin

centikelvins

decakelvin

decakelvins

decikelvin

decikelvins

degree_celsius

degrees Celsius

degree_fahrenheit

degrees Fahrenheit

degree_rankine

degrees Rankine

exakelvin

exakelvins

femtokelvin

femtokelvins

gigakelvin

gigakelvins

hectokelvin

hectokelvins

kelvin

The kelvin is the SI unit of thermodynamic temperature. It is defined by taking the fixed numerical value of the Boltzmann constant k to be 1.380 649 × 10⁻²³ when expressed in the unit J K⁻¹, which is equal to kg m² s⁻² K⁻¹, where the kilogram, meter, and second are defined in terms of h, c, and ∆νCs.

kilokelvin

kilokelvins

megakelvin

megakelvins

microkelvin

microkelvins

millikelvin

millikelvins

nanokelvin

nanokelvins

petakelvin

petakelvins

picokelvin

picokelvins

terakelvin

terakelvins

yoctokelvin

yoctokelvins

yottakelvin

yottakelvins

zeptokelvin

zeptokelvins

zettakelvin

zettakelvins

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of thermodynamic temperature, Th (base unit kelvin, K).

ThermodynamicTemperature

Thermodynamic temperature (base unit kelvin, K).