Type Definition uom::si::thermodynamic_temperature::ThermodynamicTemperature

type ThermodynamicTemperature<U, V> = Quantity<Dimension, U, V>;

Thermodynamic temperature (base unit kelvin, K1).

Methods

impl<U: ?Sized, V> ThermodynamicTemperature<U, V> where
U: Units<V>,
V: Num + Conversion<V>,

pub fn new<N>(v: V) -> Self where
N: Unit + Conversion<V, T = V::T>,

Create a new quantity from the given value and measurement unit.

pub fn get<N>(&self) -> V where
N: Unit + Conversion<V, T = V::T>,

Retrieve the value of the quantity in the given measurement unit.

pub fn floor<N>(self) -> Self where
V: Float,
N: Unit + Conversion<V, T = V::T>,

Returns the largest integer less than or equal to a number in the given measurement unit.

pub fn ceil<N>(self) -> Self where
V: Float,
N: Unit + Conversion<V, T = V::T>,

Returns the smallest integer less than or equal to a number in the given measurement unit.

pub fn round<N>(self) -> Self where
V: Float,
N: Unit + Conversion<V, T = V::T>,

Returns the nearest integer to a number in the in given measurement unit. Round half-way cases away from 0.0.

pub fn trunc<N>(self) -> Self where
V: Float,
N: Unit + Conversion<V, T = V::T>,

Returns the integer part of a number in the given measurement unit.

pub fn fract<N>(self) -> Self where
V: Float,
N: Unit + Conversion<V, T = V::T>,

Returns the fractional part of a number in the given measurement unit.