Module uom::si::velocity

Velocity (base unit meter per second, m · s⁻¹).

Structs

atomic_unit_of_velocity

Hartree atomic unit of velocity a₀ · Eₕ/ħ, where a₀ is Bohr radius, Eₕ is Hartree energy, and ħ is reduced Planck constant

attometer_per_second

attometers per second

centimeter_per_second

centimeters per second

decameter_per_second

decameters per second

decimeter_per_second

decimeters per second

exameter_per_second

exameters per second

femtometer_per_second

femtometers per second

foot_per_hour

feet per hour

foot_per_minute

feet per minute

foot_per_second

feet per second

gigameter_per_second

gigameters per second

hectometer_per_second

hectometers per second

inch_per_second

inches per second

kilometer_per_hour

kilometers per hour

kilometer_per_second

kilometers per second

knot

knots

megameter_per_second

megameters per second

meter_per_second

meters per second

micrometer_per_second

micrometers per second

mile_per_hour

miles per hour

mile_per_minute

miles per minute

mile_per_second

miles per second

millimeter_per_minute

millimeters per minute

millimeter_per_second

millimeters per second

nanometer_per_second

nanometers per second

natural_unit_of_velocity

Speed of light in vacuum, CODATA 2018

petameter_per_second

petameters per second

picometer_per_second

picometers per second

speed_of_light_in_vacuum

Speed of light in vacuum, CODATA 2018

terameter_per_second

terameters per second

yoctometer_per_second

yoctometers per second

yottameter_per_second

yottameters per second

zeptometer_per_second

zeptometers per second

zettameter_per_second

zettameters per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of velocity, LT⁻¹ (base unit meter per second, m · s⁻¹).

Velocity

Velocity (base unit meter per second, m · s⁻¹).