Module uom::si::volume

Volume (base unit cubic meter, m³).

Structs

acre_foot

acre-feet

attoliter

attoliters

barrel

barrels

bushel

bushels

centiliter

centiliters

cord

cords

cubic_attometer

cubic attometers

cubic_centimeter

cubic centimeters

cubic_decameter

cubic decameters

cubic_decimeter

cubic decimeters

cubic_exameter

cubic exameters

cubic_femtometer

cubic femtometers

cubic_foot

cubic feet

cubic_gigameter

cubic gigameters

cubic_hectometer

cubic hectometers

cubic_inch

cubic inches

cubic_kilometer

cubic kilometers

cubic_megameter

cubic megameters

cubic_meter

cubic meters

cubic_micrometer

cubic micrometers

cubic_mile

cubic miles

cubic_millimeter

cubic millimeters

cubic_nanometer

cubic nanometers

cubic_petameter

cubic petameters

cubic_picometer

cubic picometers

cubic_terameter

cubic terameters

cubic_yard

cubic yards

cubic_yoctometer

cubic yoctometers

cubic_yottameter

cubic yottameters

cubic_zeptometer

cubic zeptometers

cubic_zettameter

cubic zettameters

cup

cups

decaliter

decaliters

deciliter

deciliters

exaliter

exaliters

femtoliter

femtoliters

fluid_ounce

fluid ounces

fluid_ounce_imperial

Imperial fluid ounces

gallon

gallons

gallon_imperial

Imperial gallons

gigaliter

gigaliters

gill

gills

gill_imperial

Imperial gills

hectoliter

hectoliters

kiloliter

kiloliters

liter

liters

megaliter

megaliters

microliter

microliters

milliliter

milliliters

nanoliter

nanoliters

peck

pecks

petaliter

petaliters

picoliter

picoliters

pint_dry

dry pints

pint_liquid

liquid pints

quart_dry

dry quarts

quart_liquid

liquid quarts

register_ton

register tons

stere

steres

tablespoon

tablespoons

teaspoon

teaspoons

teraliter

teraliters

yoctoliter

yoctoliters

yottaliter

yottaliters

zeptoliter

zeptoliters

zettaliter

zettaliters

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of volume, L³ (base unit cubic meter, m³).

Volume

Volume (base unit cubic meter, m³).