Type Alias uom::si::volume::Volume

pub type Volume<U, V> = Quantity<Dimension, U, V>;

Volume (base unit cubic meter, m³).

Generic Parameters

U: Base units.

V: Underlying storage type.

Aliased Type

struct Volume<U, V> {
    pub dimension: PhantomData<dyn Dimension<I = Z0, M = Z0, N = Z0, Th = Z0, T = Z0, L = PInt<UInt<UInt<UTerm, B1>, B1>>, J = Z0, Kind = dyn Kind>>,
    pub units: PhantomData<U>,
    pub value: V,
}

Fields
dimension: PhantomData<dyn Dimension<I = Z0, M = Z0, N = Z0, Th = Z0, T = Z0, L = PInt<UInt<UInt<UTerm, B1>, B1>>, J = Z0, Kind = dyn Kind>>

Quantity dimension. See Dimension.

units: PhantomData<U>

Quantity base units. See Units.

value: V

Quantity value stored in the base units for the quantity.

Implementations

impl<U, V> Volume<U, V>
where U: Units<V> + ?Sized, V: Num + Conversion<V>,

pub fn new<N>(v: V) -> Self
where N: Unit + Conversion<V, T = V::T>,

Create a new quantity from the given value and measurement unit.

Generic Parameters

N: Unit.

pub fn get<N>(&self) -> V
where N: Unit + Conversion<V, T = V::T>,

Retrieve the value of the quantity in the given measurement unit.

Generic Parameters

N: Unit.

pub fn floor<N>(self) -> Self
where V: Float, N: Unit + Conversion<V, T = V::T>,

Returns the largest integer less than or equal to a number in the given measurement unit.

Generic Parameters

N: Unit.

pub fn ceil<N>(self) -> Self
where V: Float, N: Unit + Conversion<V, T = V::T>,

Returns the smallest integer less than or equal to a number in the given measurement unit.

Generic Parameters

N: Unit.

pub fn round<N>(self) -> Self
where V: Float, N: Unit + Conversion<V, T = V::T>,

Returns the nearest integer to a number in the in given measurement unit. Round half-way cases away from 0.0.

Generic Parameters

N: Unit.

pub fn trunc<N>(self) -> Self
where V: Float, N: Unit + Conversion<V, T = V::T>,

Returns the integer part of a number in the given measurement unit.

Generic Parameters

N: Unit.

pub fn fract<N>(self) -> Self
where V: Float, N: Unit + Conversion<V, T = V::T>,

Returns the fractional part of a number in the given measurement unit.

Generic Parameters

N: Unit.

pub fn format_args<N>(_unit: N, style: DisplayStyle) -> Arguments<Dimension, N>
where N: Unit,

Creates a struct that can be used to format a compatible quantity for display.

Notes

The return value of this method cannot be used to print directly, but is instead used to format quantities and can be reused; see Arguments::with and the examples below.

If you do not need to format multiple quantities, consider using into_format_args instead.

Examples

let t1 = Time::new::<femtosecond>(1.0_E-1);
let t2 = Time::new::<picosecond>(1.0_E-1);
let a = Time::format_args(femtosecond, Description);

assert_eq!("0.1 femtoseconds", format!("{}", a.with(t1)));
assert_eq!("100 femtoseconds", format!("{}", a.with(t2)));

Generic Parameters

N: Unit.

pub fn into_format_args<N>( self, _unit: N, style: DisplayStyle, ) -> QuantityArguments<Dimension, U, V, N>
where N: Unit,

Creates a struct that formats self for display.

Notes

Unlike format_args, the return value of this method can be used directly for display. It will format the value of self for the quantity on which it is called and nothing else.

If you wish to reuse the return value to format multiple quantities, use format_args instead.

Examples

let t = Time::new::<picosecond>(1.0_E-1);
let a = t.into_format_args(femtosecond, Description);

assert_eq!("100 femtoseconds", format!("{}", a));

Generic Parameters

N: Unit.