Module uom::si::volume_rate

Volume rate (base unit cubic meter per second, m³ · s⁻¹).

Structs

acre_foot_per_second

acre-feet per second

attoliter_per_second

attoliters per second

barrel_per_second

barrels per second

bushel_per_second

bushels per second

centiliter_per_second

centiliters per second

cord_per_second

cords per second

cubic_attometer_per_second

cubic attometers per second

cubic_centimeter_per_second

cubic centimeters per second

cubic_decameter_per_second

cubic decameters per second

cubic_decimeter_per_second

cubic decimeters per second

cubic_exameter_per_second

cubic exameters per second

cubic_femtometer_per_second

cubic femtometers per second

cubic_foot_per_minute

cubic feet per minute

cubic_foot_per_second

cubic feet per second

cubic_gigameter_per_second

cubic gigameters per second

cubic_hectometer_per_second

cubic hectometers per second

cubic_inch_per_minute

cubic inches per minute

cubic_inch_per_second

cubic inches per second

cubic_kilometer_per_second

cubic kilometers per second

cubic_megameter_per_second

cubic megameters per second

cubic_meter_per_hour

cubic meters per hour

cubic_meter_per_minute

cubic meters per minute

cubic_meter_per_second

cubic meters per second

cubic_micrometer_per_second

cubic micrometers per second

cubic_mile_per_second

cubic miles per second

cubic_millimeter_per_second

cubic millimeters per second

cubic_nanometer_per_second

cubic nanometers per second

cubic_petameter_per_second

cubic petameters per second

cubic_picometer_per_second

cubic picometers per second

cubic_terameter_per_second

cubic terameters per second

cubic_yard_per_minute

cubic yards per minute

cubic_yard_per_second

cubic yards per second

cubic_yoctometer_per_second

cubic yoctometers per second

cubic_yottameter_per_second

cubic yottameters per second

cubic_zeptometer_per_second

cubic zeptometers per second

cubic_zettameter_per_second

cubic zettameters per second

cup_per_second

cups per second

decaliter_per_second

decaliters per second

deciliter_per_second

deciliters per second

exaliter_per_second

exaliters per second

femtoliter_per_second

femtoliters per second

fluid_ounce_imperial_per_second

Imperial fluid ounces per second

fluid_ounce_per_second

fluid ounces per second

gallon_imperial_per_second

Imperial gallons per second

gallon_per_day

gallons per day

gallon_per_minute

gallons per minute

gallon_per_second

gallons per second

gigaliter_per_second

gigaliters per second

gill_imperial_per_second

Imperial gills per second

gill_per_second

gills per second

hectoliter_per_second

hectoliters per second

kiloliter_per_second

kiloliters per second

liter_per_minute

liters per minute

liter_per_second

liters per second

megaliter_per_second

megaliters per second

microliter_per_second

microliters per second

milliliter_per_second

milliliters per second

nanoliter_per_second

nanoliters per second

peck_per_second

pecks per second

petaliter_per_second

petaliters per second

picoliter_per_second

picoliters per second

pint_dry_per_second

dry pints per second

pint_liquid_per_second

liquid pints per second

quart_dry_per_second

dry quarts per second

quart_liquid_per_second

liquid quarts per second

register_ton_per_second

register tons per second

stere_per_second

steres per second

tablespoon_per_second

tablespoons per second

teaspoon_per_second

teaspoons per second

teraliter_per_second

teraliters per second

yoctoliter_per_second

yoctoliters per second

yottaliter_per_second

yottaliters per second

zeptoliter_per_second

zeptoliters per second

zettaliter_per_second

zettaliters per second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of volume rate, L³T⁻¹ (base unit cubic meter per second, m³ · s⁻¹).

VolumeRate

Volume rate (base unit cubic meter per second, m³ · s⁻¹).