Module uom::si::volumetric_number_density

Volumetric number density (base unit 1 per cubic meter, m⁻³).

Structs

per_centiliter

per centiliter

per_cubic_centimeter

per cubic centimeter

per_cubic_decimeter

per cubic decimeter

per_cubic_foot

per cubic foot

per_cubic_inch

per cubic inch

per_cubic_kilometer

per cubic kilometer

per_cubic_meter

per cubic meter

per_cubic_mile

per cubic mile

per_cubic_millimeter

per cubic millimeter

per_cubic_yard

per cubic yard

per_deciliter

per deciliter

per_fluid_ounce

per fluid ounce

per_fluid_ounce_imperial

per Imperial fluid ounce

per_gallon

per gallon

per_gallon_imperial

per Imperial gallon

per_liter

per liter

per_milliliter

per milliliter

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of volumetric number density, L⁻³ (base unit 1 per cubic meter, m⁻³).

VolumetricNumberDensity

Volumetric number density (base unit 1 per cubic meter, m⁻³).