Module uom::si::volumetric_number_rate

Volumetric number rate (base unit 1 per cubic meter second, m⁻³ · s⁻¹).

Structs

becquerel_per_cubic_meter

becquerels per cubic meter

curie_per_cubic_meter

curies per cubic meter

disintegrations_per_minute_per_cubic_meter

disintegrations per minute per cubic meter

per_cubic_centimeter_second

per cubic centimeter second

per_cubic_foot_second

per cubic foot second

per_cubic_inch_second

per cubic inch second

per_cubic_meter_second

per cubic meter second

per_cubic_mile_second

per cubic mile second

per_cubic_millimeter_second

per cubic millimeter second

per_cubic_yard_second

per cubic yard second

per_fluid_ounce_imperial_second

per Imperial fluid ounce second

per_fluid_ounce_second

per fluid ounce second

per_gallon_imperial_second

per Imperial gallon second

per_gallon_second

per gallon second

per_liter_second

per liter second

per_milliliter_second

per milliliter second

Enums

Units

Unit enum.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See crate::Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

units

Iterate over all defined units for this quantity.

Type Aliases

Dimension

Dimension of volumetric number rate, L⁻³T⁻¹ (base unit 1 per cubic meter second, m⁻³ · s⁻¹).

VolumetricNumberRate

Volumetric number rate (base unit 1 per cubic meter second, m⁻³ · s⁻¹).