Module uom::si

International System of Units (SI) and International System of Quantities (ISQ) implementations.

Modules

acceleration
Acceleration (base unit meter per second squared, m · s⁻²).

amount_of_substance
Amount of substance (base unit mole, mol).

angle
Angle (dimensionless quantity).

angular_acceleration
Angular acceleration (base unit radian per second squared, s⁻²).

angular_jerk
Angular jerk (base unit radian per second cubed, s⁻³).

angular_velocity
Angular velocity (base unit radian per second, s⁻¹).

area
Area (base unit square meter, m²).

available_energy
Available energy (base unit joule per kilogram, m² · s⁻²).

capacitance
Capacitance (base unit farad, m⁻² · kg⁻¹ · s⁴ · A²).

electric_charge
Electric charge (base unit coulomb, A · s).

electric_current
Electric current (base unit ampere, A).

electric_potential
Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).

electrical_conductance
Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

electrical_resistance
Electrical resistance (base unit ohm, m² · kg · s⁻³ · A⁻²).

energy
Energy (base unit joule, kg · m² · s⁻²).

f32
Quantity type aliases using the default base units.

fmt
Utilities for formatting and printing quantities.

force
Force (base unit newton, kg · m · s⁻²).

frequency
Frequency (base unit hertz, s⁻¹).

heat_capacity
Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

heat_flux_density
Heat flux density (base unit watt per square meter, kg · s⁻³).

heat_transfer
Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

inductance
Inductance (base unit henry, m² · kg · s⁻² · A⁻²).

information
Information (dimensionless quantity).

information_rate
Information rate (base unit byte per second, s⁻¹).

jerk
Jerk (base unit meter per second cubed, m · s⁻³).

length
Length (base unit meter, m).

luminance
Luminance (base unit candela per square meter, cd · m⁻²).

luminous_intensity
Luminous intensity (base unit candela, cd).

magnetic_flux
Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).

magnetic_flux_density
Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).

marker
Primitive traits and types representing basic properties of types specific to the SI.

mass
Mass (base unit kilogram, kg).

mass_density
Mass density (base unit kilogram per cubic meter, kg · m⁻³).

mass_rate
Mass rate (base unit kilogram per second, kg · s⁻¹).

molar_energy
Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

molar_mass
Molar mass (base unit kilogram per mole, kg · mol⁻¹).

momentum
Momentum (base unit kilogram meter per second, kg · m · s⁻¹).

power
Power (base unit watt, m² · kg · s⁻³).

pressure
Pressure (base unit pascal, kg · m⁻¹ · s⁻²).

quantities
Quantity type aliases using the default base units and parameterized on the underlying storage type.

ratio
Ratio (dimensionless quantity).

specific_heat_capacity
Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

temperature_interval
Temperature interval (base unit kelvin, K).

thermal_conductivity
Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

thermodynamic_temperature
Thermodynamic temperature (base unit kelvin, K).

time
Time (base unit second, s).

torque
Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).

velocity
Velocity (base unit meter per second, m · s⁻¹).

volume
Volume (base unit cubic meter, m³).

volume_rate
Volume rate (base unit cubic meter per second, m³ · s⁻¹).

Structs

Quantity
Property of a phenomenon, body or substance, where the property has a magnitude that can be expressed as a number and a reference.

Traits

Dimension
Marker trait to express the dependence of a quantity on the base quantities of a system of quantities as a product of powers of factors corresponding to the base quantities, omitting any numerical factor.

Unit
Trait to identify measurement units of individual quantities.

Units
Marker trait to identify a system of units based on a set of base units of a system of quantities.

Type Definitions

DimensionOne
Type alias for dimension one for which all the exponents of the factors corresponding to the base quantities are zero.

ISQ
International System of Quantities (ISQ).

SI
International System of Units (SI).