Module uom::si::acceleration

Acceleration (base unit meter per second squared, m · s⁻²).

Structs

attometer_per_second_squared
Measurement unit.

centimeter_per_second_squared
Measurement unit.

decameter_per_second_squared
Measurement unit.

decimeter_per_second_squared
Measurement unit.

exameter_per_second_squared
Measurement unit.

femtometer_per_second_squared
Measurement unit.

foot_per_second_squared
Measurement unit.

galileo
Measurement unit.

gigameter_per_second_squared
Measurement unit.

hectometer_per_second_squared
Measurement unit.

inch_per_second_squared
Measurement unit.

kilometer_per_second_squared
Measurement unit.

megameter_per_second_squared
Measurement unit.

meter_per_second_squared
Measurement unit.

micrometer_per_second_squared
Measurement unit.

millimeter_per_minute_squared
Measurement unit.

millimeter_per_second_squared
Measurement unit.

nanometer_per_second_squared
Measurement unit.

petameter_per_second_squared
Measurement unit.

picometer_per_second_squared
Measurement unit.

terameter_per_second_squared
Measurement unit.

yoctometer_per_second_squared
Measurement unit.

yottameter_per_second_squared
Measurement unit.

zeptometer_per_second_squared
Measurement unit.

zettameter_per_second_squared
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Acceleration
Acceleration (base unit meter per second squared, m · s⁻²).

Dimension
Dimension of acceleration, LT⁻² (base unit meter per second squared, m · s⁻²).