Module uom::si::amount_of_substance

Amount of substance (base unit mole, mol1).

Structs

attomole
Measurement unit.

centimole
Measurement unit.

decamole
Measurement unit.

decimole
Measurement unit.

examole
Measurement unit.

femtomole
Measurement unit.

gigamole
Measurement unit.

hectomole
Measurement unit.

kilomole
Measurement unit.

megamole
Measurement unit.

micromole
Measurement unit.

millimole
Measurement unit.

mole

The mole is the amount of substance of a system which contains as many elementary entities as there are atoms in 0.012 kilogram of carbon 12.

When the mole is used, the elementary entities must be specified and may be atoms, molecules, ions, electrons, other particles, or specified groups of such particles.

nanomole
Measurement unit.

petamole
Measurement unit.

picomole
Measurement unit.

teramole
Measurement unit.

yoctomole
Measurement unit.

yottamole
Measurement unit.

zeptomole
Measurement unit.

zettamole
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

AmountOfSubstance
Amount of substance (base unit mole, mol1).

Dimension
Amount of substance dimension, mol1.