Module uom::si::area

Area (base unit square meter, m2).

Structs

acre
Measurement unit.

are
Measurement unit.

barn
Measurement unit.

circular_mil
Measurement unit.

hectare
Measurement unit.

square_attometer
Measurement unit.

square_centimeter
Measurement unit.

square_decameter
Measurement unit.

square_decimeter
Measurement unit.

square_exameter
Measurement unit.

square_femtometer
Measurement unit.

square_foot
Measurement unit.

square_gigameter
Measurement unit.

square_hectometer
Measurement unit.

square_inch
Measurement unit.

square_kilometer
Measurement unit.

square_megameter
Measurement unit.

square_meter
Measurement unit.

square_micrometer
Measurement unit.

square_mile
Measurement unit.

square_millimeter
Measurement unit.

square_nanometer
Measurement unit.

square_petameter
Measurement unit.

square_picometer
Measurement unit.

square_terameter
Measurement unit.

square_yard
Measurement unit.

square_yoctometer
Measurement unit.

square_yottameter
Measurement unit.

square_zeptometer
Measurement unit.

square_zettameter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Area
Area (base unit square meter, m2).

Dimension
Area dimension, m2.