Module uom::si::available_energy

Available energy (base unit joule per kilogram, m² · s⁻²).

Structs

attojoule_per_kilogram
Measurement unit.

btu_it_per_pound
Measurement unit.

btu_per_pound
Measurement unit.

calorie_it_per_gram
Measurement unit.

calorie_per_gram
Measurement unit.

centijoule_per_kilogram
Measurement unit.

decajoule_per_kilogram
Measurement unit.

decijoule_per_kilogram
Measurement unit.

exajoule_per_kilogram
Measurement unit.

femtojoule_per_kilogram
Measurement unit.

gigajoule_per_kilogram
Measurement unit.

hectojoule_per_kilogram
Measurement unit.

joule_per_attogram
Measurement unit.

joule_per_exagram
Measurement unit.

joule_per_femtogram
Measurement unit.

joule_per_gigagram
Measurement unit.

joule_per_gram
Measurement unit.

joule_per_kilogram
Derived unit of available energy.

joule_per_megagram
Measurement unit.

joule_per_microgram
Measurement unit.

joule_per_milligram
Measurement unit.

joule_per_nanogram
Measurement unit.

joule_per_petagram
Measurement unit.

joule_per_picogram
Measurement unit.

joule_per_teragram
Measurement unit.

joule_per_yottagram
Measurement unit.

joule_per_zeptogram
Measurement unit.

joule_per_zettagram
Measurement unit.

kilojoule_per_kilogram
Measurement unit.

megajoule_per_kilogram
Measurement unit.

microjoule_per_kilogram
Measurement unit.

millijoule_per_kilogram
Measurement unit.

nanojoule_per_kilogram
Measurement unit.

petajoule_per_kilogram
Measurement unit.

picojoule_per_kilogram
Measurement unit.

terajoule_per_kilogram
Measurement unit.

yoctojoule_per_kilogram
Measurement unit.

yottajoule_per_kilogram
Measurement unit.

zeptojoule_per_kilogram
Measurement unit.

zettajoule_per_kilogram
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

AvailableEnergy
Available energy (base unit joule per kilogram, m² · s⁻²).

Dimension
Dimension of available energy, L²T⁻² (base unit joule per kilogram, m² · s⁻²).