Module uom::si::catalytic_activity

Catalytic activity (base unit katal, mol · s⁻¹).

Structs

atto_enzyme_unit
Measurement unit.

attokatal
Measurement unit.

centi_enzyme_unit
Measurement unit.

centikatal
Measurement unit.

deca_enzyme_unit
Measurement unit.

decakatal
Measurement unit.

deci_enzyme_unit
Measurement unit.

decikatal
Measurement unit.

enzyme_unit
Measurement unit.

exa_enzyme_unit
Measurement unit.

exakatal
Measurement unit.

femto_enzyme_unit
Measurement unit.

femtokatal
Measurement unit.

giga_enzyme_unit
Measurement unit.

gigakatal
Measurement unit.

hecto_enzyme_unit
Measurement unit.

hectokatal
Measurement unit.

katal
Measurement unit.

kilo_enzyme_unit
Measurement unit.

kilokatal
Measurement unit.

mega_enzyme_unit
Measurement unit.

megakatal
Measurement unit.

micro_enzyme_unit
Measurement unit.

microkatal
Measurement unit.

milli_enzyme_unit
Measurement unit.

millikatal
Measurement unit.

nano_enzyme_unit
Measurement unit.

nanokatal
Measurement unit.

peta_enzyme_unit
Measurement unit.

petakatal
Measurement unit.

pico_enzyme_unit
Measurement unit.

picokatal
Measurement unit.

tera_enzyme_unit
Measurement unit.

terakatal
Measurement unit.

yocto_enzyme_unit
Measurement unit.

yoctokatal
Measurement unit.

yotta_enzyme_unit
Measurement unit.

yottakatal
Measurement unit.

zepto_enzyme_unit
Measurement unit.

zeptokatal
Measurement unit.

zetta_enzyme_unit
Measurement unit.

zettakatal
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

CatalyticActivity
Catalytic activity (base unit katal, mol · s⁻¹).

Dimension
Dimension of catalytic activity, T⁻¹N (base unit katal, mol · s⁻¹).