Module uom::si::catalytic_activity_concentration

Catalytic activity concentration (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).

Structs

atto_enzyme_unit_per_cubic_meter
Measurement unit.

attokatal_per_cubic_meter
Measurement unit.

centi_enzyme_unit_per_cubic_meter
Measurement unit.

centikatal_per_cubic_meter
Measurement unit.

deca_enzyme_unit_per_cubic_meter
Measurement unit.

decakatal_per_cubic_meter
Measurement unit.

deci_enzyme_unit_per_cubic_meter
Measurement unit.

decikatal_per_cubic_meter
Measurement unit.

enzyme_unit_per_cubic_decimeter
Measurement unit.

enzyme_unit_per_cubic_meter
Measurement unit.

enzyme_unit_per_deciliter
Measurement unit.

enzyme_unit_per_liter
Measurement unit.

enzyme_unit_per_milliliter
Measurement unit.

exa_enzyme_unit_per_cubic_meter
Measurement unit.

exakatal_per_cubic_meter
Measurement unit.

femto_enzyme_unit_per_cubic_decimeter
Measurement unit.

femto_enzyme_unit_per_cubic_meter
Measurement unit.

femto_enzyme_unit_per_deciliter
Measurement unit.

femto_enzyme_unit_per_liter
Measurement unit.

femto_enzyme_unit_per_milliliter
Measurement unit.

femtokatal_per_cubic_decimeter
Measurement unit.

femtokatal_per_cubic_meter
Measurement unit.

femtokatal_per_deciliter
Measurement unit.

femtokatal_per_liter
Measurement unit.

femtokatal_per_milliliter
Measurement unit.

giga_enzyme_unit_per_cubic_meter
Measurement unit.

gigakatal_per_cubic_meter
Measurement unit.

hecto_enzyme_unit_per_cubic_meter
Measurement unit.

hectokatal_per_cubic_meter
Measurement unit.

katal_per_cubic_decimeter
Measurement unit.

katal_per_cubic_meter
Measurement unit.

katal_per_deciliter
Measurement unit.

katal_per_liter
Measurement unit.

katal_per_milliliter
Measurement unit.

kilo_enzyme_unit_per_cubic_decimeter
Measurement unit.

kilo_enzyme_unit_per_cubic_meter
Measurement unit.

kilo_enzyme_unit_per_deciliter
Measurement unit.

kilo_enzyme_unit_per_liter
Measurement unit.

kilo_enzyme_unit_per_milliliter
Measurement unit.

kilokatal_per_cubic_decimeter
Measurement unit.

kilokatal_per_cubic_meter
Measurement unit.

kilokatal_per_deciliter
Measurement unit.

kilokatal_per_liter
Measurement unit.

kilokatal_per_milliliter
Measurement unit.

mega_enzyme_unit_per_cubic_meter
Measurement unit.

megakatal_per_cubic_meter
Measurement unit.

micro_enzyme_unit_per_cubic_decimeter
Measurement unit.

micro_enzyme_unit_per_cubic_meter
Measurement unit.

micro_enzyme_unit_per_deciliter
Measurement unit.

micro_enzyme_unit_per_liter
Measurement unit.

micro_enzyme_unit_per_milliliter
Measurement unit.

microkatal_per_cubic_decimeter
Measurement unit.

microkatal_per_cubic_meter
Measurement unit.

microkatal_per_deciliter
Measurement unit.

microkatal_per_liter
Measurement unit.

microkatal_per_milliliter
Measurement unit.

milli_enzyme_unit_per_cubic_decimeter
Measurement unit.

milli_enzyme_unit_per_cubic_meter
Measurement unit.

milli_enzyme_unit_per_deciliter
Measurement unit.

milli_enzyme_unit_per_liter
Measurement unit.

milli_enzyme_unit_per_milliliter
Measurement unit.

millikatal_per_cubic_decimeter
Measurement unit.

millikatal_per_cubic_meter
Measurement unit.

millikatal_per_deciliter
Measurement unit.

millikatal_per_liter
Measurement unit.

millikatal_per_milliliter
Measurement unit.

nano_enzyme_unit_per_cubic_decimeter
Measurement unit.

nano_enzyme_unit_per_cubic_meter
Measurement unit.

nano_enzyme_unit_per_deciliter
Measurement unit.

nano_enzyme_unit_per_liter
Measurement unit.

nano_enzyme_unit_per_milliliter
Measurement unit.

nanokatal_per_cubic_decimeter
Measurement unit.

nanokatal_per_cubic_meter
Measurement unit.

nanokatal_per_deciliter
Measurement unit.

nanokatal_per_liter
Measurement unit.

nanokatal_per_milliliter
Measurement unit.

peta_enzyme_unit_per_cubic_meter
Measurement unit.

petakatal_per_cubic_meter
Measurement unit.

pico_enzyme_unit_per_cubic_decimeter
Measurement unit.

pico_enzyme_unit_per_cubic_meter
Measurement unit.

pico_enzyme_unit_per_deciliter
Measurement unit.

pico_enzyme_unit_per_liter
Measurement unit.

pico_enzyme_unit_per_milliliter
Measurement unit.

picokatal_per_cubic_decimeter
Measurement unit.

picokatal_per_cubic_meter
Measurement unit.

picokatal_per_deciliter
Measurement unit.

picokatal_per_liter
Measurement unit.

picokatal_per_milliliter
Measurement unit.

tera_enzyme_unit_per_cubic_meter
Measurement unit.

terakatal_per_cubic_meter
Measurement unit.

yocto_enzyme_unit_per_cubic_meter
Measurement unit.

yoctokatal_per_cubic_meter
Measurement unit.

yotta_enzyme_unit_per_cubic_meter
Measurement unit.

yottakatal_per_cubic_meter
Measurement unit.

zepto_enzyme_unit_per_cubic_meter
Measurement unit.

zeptokatal_per_cubic_meter
Measurement unit.

zetta_enzyme_unit_per_cubic_meter
Measurement unit.

zettakatal_per_cubic_meter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

CatalyticActivityConcentration
Catalytic activity concentration (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).

Dimension
Dimension of catalytic activity concentration, L⁻³T⁻¹N (base unit katal per cubic meter, mol · s⁻¹ · m⁻³).