Module uom::si::electric_charge

Electric charge (base unit coulomb, A · s).

Structs

abcoulomb
Measurement unit.

ampere_hour
Measurement unit.

attocoulomb
Measurement unit.

centicoulomb
Measurement unit.

coulomb
Derived unit of electric charge.

decaampere_hour
Measurement unit.

decacoulomb
Measurement unit.

decicoulomb
Measurement unit.

exacoulomb
Measurement unit.

faraday
Measurement unit.

femtocoulomb
Measurement unit.

franklin
Measurement unit.

gigaampere_hour
Measurement unit.

gigacoulomb
Measurement unit.

hectoampere_hour
Measurement unit.

hectocoulomb
Measurement unit.

kiloampere_hour
Measurement unit.

kilocoulomb
Measurement unit.

megaampere_hour
Measurement unit.

megacoulomb
Measurement unit.

microampere_hour
Measurement unit.

microcoulomb
Measurement unit.

milliampere_hour
Measurement unit.

millicoulomb
Measurement unit.

nanocoulomb
Measurement unit.

petaampere_hour
Measurement unit.

petacoulomb
Measurement unit.

picocoulomb
Measurement unit.

statcoulomb
Measurement unit.

teraampere_hour
Measurement unit.

teracoulomb
Measurement unit.

yoctocoulomb
Measurement unit.

yottacoulomb
Measurement unit.

zeptocoulomb
Measurement unit.

zettacoulomb
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of electric charge, TI (base unit coulomb, A · s).

ElectricCharge
Electric charge (base unit coulomb, A · s).