Module uom::si::electric_charge

Electric charge (base unit coulomb, A · s).

Structs

abcoulomb
Measurement unit.

ampere_hour
Measurement unit.

attocoulomb
Measurement unit.

centicoulomb
Measurement unit.

coulomb
Derived unit of electric charge.

decacoulomb
Measurement unit.

decicoulomb
Measurement unit.

exacoulomb
Measurement unit.

faraday
Measurement unit.

femtocoulomb
Measurement unit.

franklin
Measurement unit.

gigacoulomb
Measurement unit.

hectocoulomb
Measurement unit.

kilocoulomb
Measurement unit.

megacoulomb
Measurement unit.

microcoulomb
Measurement unit.

millicoulomb
Measurement unit.

nanocoulomb
Measurement unit.

petacoulomb
Measurement unit.

picocoulomb
Measurement unit.

statcoulomb
Measurement unit.

teracoulomb
Measurement unit.

yoctocoulomb
Measurement unit.

yottacoulomb
Measurement unit.

zeptocoulomb
Measurement unit.

zettacoulomb
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Electric charge dimension, A · s.

ElectricCharge
Electric charge (base unit coulomb, A · s).