Module uom::si::electric_current

Electric current (base unit ampere, A).

Structs

abampere

Measurement unit.

ampere

The ampere is the SI unit of electric current. It is defined by taking the fixed numerical value of the elementary charge e to be 1.602 176 634 × 10⁻¹⁹ when expressed in the unit C, which is equal to A s, where the second is defined in terms of ∆νCs.

attoampere

Measurement unit.

centiampere

Measurement unit.

decaampere

Measurement unit.

deciampere

Measurement unit.

exaampere

Measurement unit.

femtoampere

Measurement unit.

gigaampere

Measurement unit.

gilbert

Measurement unit.

hectoampere

Measurement unit.

kiloampere

Measurement unit.

megaampere

Measurement unit.

microampere

Measurement unit.

milliampere

Measurement unit.

nanoampere

Measurement unit.

petaampere

Measurement unit.

picoampere

Measurement unit.

statampere

Measurement unit.

teraampere

Measurement unit.

yoctoampere

Measurement unit.

yottaampere

Measurement unit.

zeptoampere

Measurement unit.

zettaampere

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of electric current, I (base unit ampere, A).

ElectricCurrent

Electric current (base unit ampere, A).