Module uom::si::electric_potential

Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).

Structs

abvolt

Measurement unit.

attovolt

Measurement unit.

centivolt

Measurement unit.

decavolt

Measurement unit.

decivolt

Measurement unit.

exavolt

Measurement unit.

femtovolt

Measurement unit.

gigavolt

Measurement unit.

hectovolt

Measurement unit.

kilovolt

Measurement unit.

megavolt

Measurement unit.

microvolt

Measurement unit.

millivolt

Measurement unit.

nanovolt

Measurement unit.

petavolt

Measurement unit.

picovolt

Measurement unit.

statvolt

Measurement unit.

teravolt

Measurement unit.

volt

Derived unit of electric potential.

yoctovolt

Measurement unit.

yottavolt

Measurement unit.

zeptovolt

Measurement unit.

zettavolt

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of electric potential, L²MT⁻³I⁻¹ (base unit volt, m² · kg · s⁻³ · A⁻¹).

ElectricPotential

Electric potential (base unit volt, m² · kg · s⁻³ · A⁻¹).