Module uom::si::electrical_conductance

Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

Structs

abmho
Measurement unit.

absiemens
Measurement unit.

attosiemens
Measurement unit.

centisiemens
Measurement unit.

decasiemens
Measurement unit.

decisiemens
Measurement unit.

exasiemens
Measurement unit.

femtosiemens
Measurement unit.

gigasiemens
Measurement unit.

hectosiemens
Measurement unit.

kilosiemens
Measurement unit.

megasiemens
Measurement unit.

mho
Measurement unit.

microsiemens
Measurement unit.

millisiemens
Measurement unit.

nanosiemens
Measurement unit.

petasiemens
Measurement unit.

picosiemens
Measurement unit.

siemens
Derived unit of electrical conductance.

statmho
Measurement unit.

statsiemens
Measurement unit.

terasiemens
Measurement unit.

yoctosiemens
Measurement unit.

yottasiemens
Measurement unit.

zeptosiemens
Measurement unit.

zettasiemens
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of electrical conductance, L⁻²M⁻¹T³I² (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).

ElectricalConductance
Electrical conductance (base unit siemens, m⁻² · kg⁻¹ · s³ · A²).