Module uom::si::energy

Energy (base unit joule, kg · m² · s⁻²).

Structs

attojoule

Measurement unit.

btu

Measurement unit.

btu_39

Measurement unit.

btu_59

Measurement unit.

btu_60

Measurement unit.

btu_it

Measurement unit.

calorie

Measurement unit.

calorie_15

Measurement unit.

calorie_20

Measurement unit.

calorie_it

Measurement unit.

calorie_it_nutrition

Measurement unit.

calorie_nutrition

Measurement unit.

centijoule

Measurement unit.

decaelectronvolt

Measurement unit.

decajoule

Measurement unit.

decawatt_hour

Measurement unit.

decijoule

Measurement unit.

electronvolt

Measurement unit.

erg

Measurement unit.

exajoule

Measurement unit.

femtojoule

Measurement unit.

foot_pound

Measurement unit.

foot_poundal

Measurement unit.

gigaelectronvolt

Measurement unit.

gigajoule

Measurement unit.

gigawatt_hour

Measurement unit.

hectoelectronvolt

Measurement unit.

hectojoule

Measurement unit.

hectowatt_hour

Measurement unit.

joule

Derived unit of energy.

kilocalorie

Measurement unit.

kilocalorie_it

Measurement unit.

kiloelectronvolt

Measurement unit.

kilojoule

Measurement unit.

kilowatt_hour

Measurement unit.

megaelectronvolt

Measurement unit.

megajoule

Measurement unit.

megawatt_hour

Measurement unit.

microjoule

Measurement unit.

microwatt_hour

Measurement unit.

millijoule

Measurement unit.

milliwatt_hour

Measurement unit.

nanojoule

Measurement unit.

petaelectronvolt

Measurement unit.

petajoule

Measurement unit.

petawatt_hour

Measurement unit.

picojoule

Measurement unit.

quad

Measurement unit.

teraelectronvolt

Measurement unit.

terajoule

Measurement unit.

terawatt_hour

Measurement unit.

therm_ec

Measurement unit.

therm_us

Measurement unit.

ton_tnt

Measurement unit.

watt_hour

Measurement unit.

watt_second

Measurement unit.

yoctojoule

Measurement unit.

yottajoule

Measurement unit.

zeptojoule

Measurement unit.

zettajoule

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of energy, L²MT⁻² (base unit joule, kg · m² · s⁻²).

Energy

Energy (base unit joule, kg · m² · s⁻²).