Module uom::si::force

Force (base unit newton, kg · m · s⁻²).

Structs

attonewton
Measurement unit.

centinewton
Measurement unit.

decanewton
Measurement unit.

decinewton
Measurement unit.

dyne
Measurement unit.

exanewton
Measurement unit.

femtonewton
Measurement unit.

giganewton
Measurement unit.

hectonewton
Measurement unit.

kilogram_force
Measurement unit.

kilonewton
Measurement unit.

kip
Measurement unit.

meganewton
Measurement unit.

micronewton
Measurement unit.

millinewton
Measurement unit.

nanonewton
Measurement unit.

newton
Derived unit of force.

ounce_force
Measurement unit.

petanewton
Measurement unit.

piconewton
Measurement unit.

pound_force
Measurement unit.

poundal
Measurement unit.

teranewton
Measurement unit.

ton_force
Measurement unit.

yoctonewton
Measurement unit.

yottanewton
Measurement unit.

zeptonewton
Measurement unit.

zettanewton
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of force, LMT⁻² (base unit newton, kg · m · s⁻²).

Force
Force (base unit newton, kg · m · s⁻²).