Module uom::si::frequency

Frequency (base unit hertz, s⁻¹).

Structs

attohertz
Measurement unit.

centihertz
Measurement unit.

cycle_per_day
Measurement unit.

cycle_per_hour
Measurement unit.

cycle_per_minute
Measurement unit.

cycle_per_shake
Measurement unit.

cycle_per_year
Measurement unit.

decahertz
Measurement unit.

decihertz
Measurement unit.

exahertz
Measurement unit.

femtohertz
Measurement unit.

gigahertz
Measurement unit.

hectohertz
Measurement unit.

hertz
The hertz is one cycle per second.

kilohertz
Measurement unit.

megahertz
Measurement unit.

microhertz
Measurement unit.

millihertz
Measurement unit.

nanohertz
Measurement unit.

petahertz
Measurement unit.

picohertz
Measurement unit.

terahertz
Measurement unit.

yoctohertz
Measurement unit.

yottahertz
Measurement unit.

zeptohertz
Measurement unit.

zettahertz
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of frequency, T⁻¹ (base unit hertz, s⁻¹).

Frequency
Frequency (base unit hertz, s⁻¹).