Module uom::si::heat_capacity

Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

Heat capacity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

This quantity might be used to define the heat capacity of an object. To define heat capacity of a material, use specific heat capacity.

Structs

attogram_square_meter_per_second_squared_kelvin
Measurement unit.

attojoule_per_kelvin
Measurement unit.

btu_it_per_degree_fahrenheit
Measurement unit.

btu_per_degree_fahrenheit
Measurement unit.

centigram_square_meter_per_second_squared_kelvin
Measurement unit.

centijoule_per_kelvin
Measurement unit.

decagram_square_meter_per_second_squared_kelvin
Measurement unit.

decajoule_per_kelvin
Measurement unit.

decigram_square_meter_per_second_squared_kelvin
Measurement unit.

decijoule_per_kelvin
Measurement unit.

exagram_square_meter_per_second_squared_kelvin
Measurement unit.

exajoule_per_kelvin
Measurement unit.

femtogram_square_meter_per_second_squared_kelvin
Measurement unit.

femtojoule_per_kelvin
Measurement unit.

gigagram_square_meter_per_second_squared_kelvin
Measurement unit.

gigajoule_per_kelvin
Measurement unit.

gram_square_meter_per_second_squared_kelvin
Measurement unit.

hectogram_square_meter_per_second_squared_kelvin
Measurement unit.

hectojoule_per_kelvin
Measurement unit.

joule_per_degree_celsius
Measurement unit.

joule_per_kelvin
Derived unit of heat capacity expressed in terms of derived unit Joule. Equivalent to kg · m²/(s² · K).

kilogram_square_meter_per_second_squared_kelvin
The derived unit of heat capacity expressed in base units. Equivalent to J/K.

kilojoule_per_degree_celsius
Measurement unit.

kilojoule_per_kelvin
Measurement unit.

megagram_square_meter_per_second_squared_kelvin
Measurement unit.

megajoule_per_kelvin
Measurement unit.

microgram_square_meter_per_second_squared_kelvin
Measurement unit.

microjoule_per_kelvin
Measurement unit.

milligram_square_meter_per_second_squared_kelvin
Measurement unit.

millijoule_per_degree_celsius
Measurement unit.

millijoule_per_kelvin
Measurement unit.

nanogram_square_meter_per_second_squared_kelvin
Measurement unit.

nanojoule_per_kelvin
Measurement unit.

petagram_square_meter_per_second_squared_kelvin
Measurement unit.

petajoule_per_kelvin
Measurement unit.

picogram_square_meter_per_second_squared_kelvin
Measurement unit.

picojoule_per_kelvin
Measurement unit.

teragram_square_meter_per_second_squared_kelvin
Measurement unit.

terajoule_per_kelvin
Measurement unit.

yoctogram_square_meter_per_second_squared_kelvin
Measurement unit.

yoctojoule_per_kelvin
Measurement unit.

yottagram_square_meter_per_second_squared_kelvin
Measurement unit.

yottajoule_per_kelvin
Measurement unit.

zeptogram_square_meter_per_second_squared_kelvin
Measurement unit.

zeptojoule_per_kelvin
Measurement unit.

zettagram_square_meter_per_second_squared_kelvin
Measurement unit.

zettajoule_per_kelvin
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of heat capacity, L²MT⁻²Th⁻¹ (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).

HeatCapacity
Heat capacity (base unit joule per kelvin, kg · m² · s⁻² · K⁻¹).