Module uom::si::heat_transfer

Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

Heat transfer is also known as heat transfer coefficient, film coefficient, or film effectiveness. Commonly expressed using derived units power and area: watt per square meter kelvin.

Heat transfer has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attogram_per_second_cubed_kelvin

Measurement unit.

attowatt_per_square_meter_kelvin

Measurement unit.

btu_it_per_hour_square_foot_degree_fahrenheit

Measurement unit.

btu_it_per_hour_square_inch_degree_fahrenheit

Measurement unit.

btu_it_per_minute_square_foot_degree_fahrenheit

Measurement unit.

btu_it_per_minute_square_inch_degree_fahrenheit

Measurement unit.

btu_it_per_second_square_foot_degree_fahrenheit

Measurement unit.

btu_it_per_second_square_inch_degree_fahrenheit

Measurement unit.

btu_per_hour_square_foot_degree_fahrenheit

Measurement unit.

btu_per_hour_square_inch_degree_fahrenheit

Measurement unit.

btu_per_minute_square_foot_degree_fahrenheit

Measurement unit.

btu_per_minute_square_inch_degree_fahrenheit

Measurement unit.

btu_per_second_square_foot_degree_fahrenheit

Measurement unit.

btu_per_second_square_inch_degree_fahrenheit

Measurement unit.

centigram_per_second_cubed_kelvin

Measurement unit.

centiwatt_per_square_meter_kelvin

Measurement unit.

decagram_per_second_cubed_kelvin

Measurement unit.

decawatt_per_square_meter_kelvin

Measurement unit.

decigram_per_second_cubed_kelvin

Measurement unit.

deciwatt_per_square_meter_kelvin

Measurement unit.

exagram_per_second_cubed_kelvin

Measurement unit.

exawatt_per_square_meter_kelvin

Measurement unit.

femtogram_per_second_cubed_kelvin

Measurement unit.

femtowatt_per_square_meter_kelvin

Measurement unit.

gigagram_per_second_cubed_kelvin

Measurement unit.

gigawatt_per_square_meter_kelvin

Measurement unit.

gram_per_second_cubed_kelvin

Measurement unit.

hectogram_per_second_cubed_kelvin

Measurement unit.

hectowatt_per_square_meter_kelvin

Measurement unit.

joule_per_second_square_meter_degree_celsius

Measurement unit.

joule_per_second_square_meter_kelvin

Measurement unit.

kilogram_per_second_cubed_degree_celsius

Measurement unit.

kilogram_per_second_cubed_kelvin

Derived unit of heat transfer in base units. Equivalent to W/(m² · K).

kilowatt_per_square_meter_kelvin

Measurement unit.

megagram_per_second_cubed_kelvin

Measurement unit.

megawatt_per_square_meter_kelvin

Measurement unit.

microgram_per_second_cubed_kelvin

Measurement unit.

microwatt_per_square_meter_kelvin

Measurement unit.

milligram_per_second_cubed_kelvin

Measurement unit.

milliwatt_per_square_meter_kelvin

Measurement unit.

nanogram_per_second_cubed_kelvin

Measurement unit.

nanowatt_per_square_meter_kelvin

Measurement unit.

petagram_per_second_cubed_kelvin

Measurement unit.

petawatt_per_square_meter_kelvin

Measurement unit.

picogram_per_second_cubed_kelvin

Measurement unit.

picowatt_per_square_meter_kelvin

Measurement unit.

teragram_per_second_cubed_kelvin

Measurement unit.

terawatt_per_square_meter_kelvin

Measurement unit.

watt_per_square_centimeter_degree_celsius

Measurement unit.

watt_per_square_kilometer_degree_celsius

Measurement unit.

watt_per_square_meter_degree_celsius

Measurement unit.

watt_per_square_meter_kelvin

Derived unit of heat transfer in derived units. Equivalent to kg/(s³ · K).

watt_per_square_millimeter_degree_celsius

Measurement unit.

yoctogram_per_second_cubed_kelvin

Measurement unit.

yoctowatt_per_square_meter_kelvin

Measurement unit.

yottagram_per_second_cubed_kelvin

Measurement unit.

yottawatt_per_square_meter_kelvin

Measurement unit.

zeptogram_per_second_cubed_kelvin

Measurement unit.

zeptowatt_per_square_meter_kelvin

Measurement unit.

zettagram_per_second_cubed_kelvin

Measurement unit.

zettawatt_per_square_meter_kelvin

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of heat transfer, MT⁻³Th⁻¹ (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).

HeatTransfer

Heat transfer (base unit watt per square meter kelvin, kg · s⁻³ · K⁻¹).