Module uom::si::inductance

Inductance (base unit henry, m² · kg · s⁻² · A⁻²).

Structs

abhenry
Measurement unit.

attohenry
Measurement unit.

centihenry
Measurement unit.

decahenry
Measurement unit.

decihenry
Measurement unit.

exahenry
Measurement unit.

femtohenry
Measurement unit.

gigahenry
Measurement unit.

hectohenry
Measurement unit.

henry
Derived unit of inductance.

kilohenry
Measurement unit.

megahenry
Measurement unit.

microhenry
Measurement unit.

millihenry
Measurement unit.

nanohenry
Measurement unit.

petahenry
Measurement unit.

picohenry
Measurement unit.

stathenry
Measurement unit.

terahenry
Measurement unit.

yoctohenry
Measurement unit.

yottahenry
Measurement unit.

zeptohenry
Measurement unit.

zettahenry
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of inductance, L²MT⁻²I⁻² (base unit henry, m² · kg · s⁻² · A⁻²).

Inductance
Inductance (base unit henry, m² · kg · s⁻² · A⁻²).