Module uom::si::information

Information (dimensionless quantity).

Structs

bit

The bit is the basic unit of information.

byte

The byte is a unit of information consisting of 8 bits.

crumb

Measurement unit.

deciban

Measurement unit.

exabit

Measurement unit.

exabyte

Measurement unit.

exbibit

Measurement unit.

exbibyte

Measurement unit.

gibibit

Measurement unit.

gibibyte

Measurement unit.

gigabit

Measurement unit.

gigabyte

Measurement unit.

hartley

Measurement unit.

kibibit

Measurement unit.

kibibyte

Measurement unit.

kilobit

Measurement unit.

kilobyte

Measurement unit.

mebibit

Measurement unit.

mebibyte

Measurement unit.

megabit

Measurement unit.

megabyte

Measurement unit.

natural_unit_of_information

Measurement unit.

nibble

Measurement unit.

octet

Measurement unit.

pebibit

Measurement unit.

pebibyte

Measurement unit.

petabit

Measurement unit.

petabyte

Measurement unit.

shannon

Measurement unit.

tebibit

Measurement unit.

tebibyte

Measurement unit.

terabit

Measurement unit.

terabyte

Measurement unit.

trit

Measurement unit.

yobibit

Measurement unit.

yobibyte

Measurement unit.

yottabit

Measurement unit.

yottabyte

Measurement unit.

zebibit

Measurement unit.

zebibyte

Measurement unit.

zettabit

Measurement unit.

zettabyte

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of information, 1 (dimensionless).

Information

Information (dimensionless quantity).