Module uom::si::information_rate

Information rate (base unit byte per second, s⁻¹).

Structs

bit_per_second

Measurement unit.

byte_per_second

Measurement unit.

exabit_per_second

Measurement unit.

exabyte_per_second

Measurement unit.

exbibit_per_second

Measurement unit.

exbibyte_per_second

Measurement unit.

gibibit_per_second

Measurement unit.

gibibyte_per_second

Measurement unit.

gigabit_per_second

Measurement unit.

gigabyte_per_second

Measurement unit.

kibibit_per_second

Measurement unit.

kibibyte_per_second

Measurement unit.

kilobit_per_second

Measurement unit.

kilobyte_per_second

Measurement unit.

mebibit_per_second

Measurement unit.

mebibyte_per_second

Measurement unit.

megabit_per_second

Measurement unit.

megabyte_per_second

Measurement unit.

octet_per_second

Measurement unit.

pebibit_per_second

Measurement unit.

pebibyte_per_second

Measurement unit.

petabit_per_second

Measurement unit.

petabyte_per_second

Measurement unit.

tebibit_per_second

Measurement unit.

tebibyte_per_second

Measurement unit.

terabit_per_second

Measurement unit.

terabyte_per_second

Measurement unit.

yobibit_per_second

Measurement unit.

yobibyte_per_second

Measurement unit.

yottabit_per_second

Measurement unit.

yottabyte_per_second

Measurement unit.

zebibit_per_second

Measurement unit.

zebibyte_per_second

Measurement unit.

zettabit_per_second

Measurement unit.

zettabyte_per_second

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of information rate, T⁻¹ (base unit byte per second, s⁻¹).

InformationRate

Information rate (base unit byte per second, s⁻¹).