Module uom::si::jerk

Jerk (base unit meter per second cubed, m · s⁻³).

Structs

attometer_per_second_cubed
Measurement unit.

centimeter_per_second_cubed
Measurement unit.

decameter_per_second_cubed
Measurement unit.

decimeter_per_second_cubed
Measurement unit.

exameter_per_second_cubed
Measurement unit.

femtometer_per_second_cubed
Measurement unit.

foot_per_second_cubed
Measurement unit.

gigameter_per_second_cubed
Measurement unit.

hectometer_per_second_cubed
Measurement unit.

inch_per_second_cubed
Measurement unit.

kilometer_per_minute_cubed
Measurement unit.

kilometer_per_second_cubed
Measurement unit.

megameter_per_second_cubed
Measurement unit.

meter_per_second_cubed
Measurement unit.

micrometer_per_second_cubed
Measurement unit.

millimeter_per_second_cubed
Measurement unit.

nanometer_per_second_cubed
Measurement unit.

petameter_per_second_cubed
Measurement unit.

picometer_per_second_cubed
Measurement unit.

terameter_per_second_cubed
Measurement unit.

yoctometer_per_second_cubed
Measurement unit.

yottameter_per_second_cubed
Measurement unit.

zeptometer_per_second_cubed
Measurement unit.

zettameter_per_second_cubed
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of jerk, LT⁻³ (base unit meter per second cubed, m · s⁻³).

Jerk
Jerk (base unit meter per second cubed, m · s⁻³).