Module uom::si::length

Length (base unit meter, m1).

Structs

angstrom
Measurement unit.

astronomical_unit
Measurement unit.

attometer
Measurement unit.

centimeter
Measurement unit.

chain
Measurement unit.

decameter
Measurement unit.

decimeter
Measurement unit.

exameter
Measurement unit.

fathom
Measurement unit.

femtometer
Measurement unit.

fermi
Measurement unit.

foot
Measurement unit.

foot_survey
Measurement unit.

gigameter
Measurement unit.

hectometer
Measurement unit.

inch
Measurement unit.

kilometer
Measurement unit.

light_year
Measurement unit.

megameter
Measurement unit.

meter
The meter is the length of the path traveled by light in a vacuum during a time interval of 1/299 792 458 of a second.

microinch
Measurement unit.

micrometer
Measurement unit.

micron
Measurement unit.

mil
Measurement unit.

mile
Measurement unit.

mile_survey
Measurement unit.

millimeter
Measurement unit.

nanometer
Measurement unit.

nautical_mile
Measurement unit.

parsec
Measurement unit.

petameter
Measurement unit.

pica_computer
Measurement unit.

pica_printers
Measurement unit.

picometer
Measurement unit.

point_computer
Measurement unit.

point_printers
Measurement unit.

rod
Measurement unit.

terameter
Measurement unit.

yard
Measurement unit.

yoctometer
Measurement unit.

yottameter
Measurement unit.

zeptometer
Measurement unit.

zettameter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Length dimension, m1.

Length
Length (base unit meter, m1).