Module uom::si::luminance

Luminance (base unit candela per square meter, cd · m⁻²).

Structs

attocandela_per_square_meter
Measurement unit.

candela_per_square_centimeter
Measurement unit.

candela_per_square_gigameter
Measurement unit.

candela_per_square_inch
Measurement unit.

candela_per_square_kilometer
Measurement unit.

candela_per_square_megameter
Measurement unit.

candela_per_square_meter
Derived unit of luminance.

candela_per_square_micrometer
Measurement unit.

candela_per_square_millimeter
Measurement unit.

candela_per_square_nanometer
Measurement unit.

candela_per_square_picometer
Measurement unit.

candela_per_square_terameter
Measurement unit.

centicandela_per_square_meter
Measurement unit.

decacandela_per_square_meter
Measurement unit.

decicandela_per_square_meter
Measurement unit.

exacandela_per_square_meter
Measurement unit.

femtocandela_per_square_meter
Measurement unit.

footlambert
Measurement unit.

gigacandela_per_square_meter
Measurement unit.

hectocandela_per_square_meter
Measurement unit.

kilocandela_per_square_meter
Measurement unit.

lambert
Measurement unit.

megacandela_per_square_meter
Measurement unit.

microcandela_per_square_meter
Measurement unit.

millicandela_per_square_meter
Measurement unit.

nanocandela_per_square_meter
Measurement unit.

petacandela_per_square_meter
Measurement unit.

picocandela_per_square_meter
Measurement unit.

stilb
Measurement unit.

teracandela_per_square_meter
Measurement unit.

yoctocandela_per_square_meter
Measurement unit.

yottacandela_per_square_meter
Measurement unit.

zeptocandela_per_square_meter
Measurement unit.

zettacandela_per_square_meter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of luminance, L⁻²J (base unit candela per square meter, cd · m⁻²).

Luminance
Luminance (base unit candela per square meter, cd · m⁻²).