Module uom::si::luminous_intensity

Luminous intensity (base unit candela, cd).

Structs

attocandela
Measurement unit.

candela
The candela is the SI unit of luminous intensity in a given direction. It is defined by taking the fixed numerical value of the luminous efficacy of monochromatic radiation of frequency 540 × 10¹² Hz, Kcd, to be 683 when expressed in the unit lm W⁻¹, which is equal to cd sr W⁻¹, or cd sr kg⁻¹ m⁻² s³, where the kilogram, meter, and second are defined in terms of h, c and ∆νCs.

centicandela
Measurement unit.

decacandela
Measurement unit.

decicandela
Measurement unit.

exacandela
Measurement unit.

femtocandela
Measurement unit.

gigacandela
Measurement unit.

hectocandela
Measurement unit.

kilocandela
Measurement unit.

megacandela
Measurement unit.

microcandela
Measurement unit.

millicandela
Measurement unit.

nanocandela
Measurement unit.

petacandela
Measurement unit.

picocandela
Measurement unit.

teracandela
Measurement unit.

yoctocandela
Measurement unit.

yottacandela
Measurement unit.

zeptocandela
Measurement unit.

zettacandela
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of luminous intensity, J (base unit candela, cd).

LuminousIntensity
Luminous intensity (base unit candela, cd).