Module uom::si::magnetic_flux

Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).

Structs

attoweber
Measurement unit.

centiweber
Measurement unit.

decaweber
Measurement unit.

deciweber
Measurement unit.

exaweber
Measurement unit.

femtoweber
Measurement unit.

gigaweber
Measurement unit.

hectoweber
Measurement unit.

kiloweber
Measurement unit.

maxwell
Measurement unit.

megaweber
Measurement unit.

microweber
Measurement unit.

milliweber
Measurement unit.

nanoweber
Measurement unit.

petaweber
Measurement unit.

picoweber
Measurement unit.

teraweber
Measurement unit.

weber
Derived unit of magnetic flux.

yoctoweber
Measurement unit.

yottaweber
Measurement unit.

zeptoweber
Measurement unit.

zettaweber
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of magnetic flux, L²MT⁻²I⁻¹ (base unit weber, m² · kg · s⁻² · A⁻¹).

MagneticFlux
Magnetic flux (base unit weber, m² · kg · s⁻² · A⁻¹).