Module uom::si::magnetic_flux_density

Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).

Structs

attotesla
Measurement unit.

centitesla
Measurement unit.

decatesla
Measurement unit.

decitesla
Measurement unit.

exatesla
Measurement unit.

femtotesla
Measurement unit.

gamma
Measurement unit.

gauss
Measurement unit.

gigatesla
Measurement unit.

hectotesla
Measurement unit.

kilotesla
Measurement unit.

megatesla
Measurement unit.

microtesla
Measurement unit.

millitesla
Measurement unit.

nanotesla
Measurement unit.

petatesla
Measurement unit.

picotesla
Measurement unit.

teratesla
Measurement unit.

tesla
Derived unit of magnetic flux density.

yoctotesla
Measurement unit.

yottatesla
Measurement unit.

zeptotesla
Measurement unit.

zettatesla
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of magnetic flux density, MT⁻²I⁻¹ (base unit tesla, kg · s⁻² · A⁻¹).

MagneticFluxDensity
Magnetic flux density (base unit tesla, kg · s⁻² · A⁻¹).