Module uom::si::marker

Primitive traits and types representing basic properties of types specific to the SI.

Traits

AngleKind
AngleKind is a Kind for separating angular quantities from their identically dimensioned non-angular quantity counterparts. Conversions to and from AngleKind quantities are supported through implementations of the From trait.

InformationKind
InformationKind is a Kind for separating information quantities from their identically dimensioned non-information quantity counterparts. Conversions to and from InformationKind quantities are supported through implementations of the From trait.

TemperatureKind
Kind of thermodynamic temperature.