Module uom::si::marker

Primitive traits and types representing basic properties of types specific to the SI.

Traits

AngleKind
AngleKind is a Kind for separating angular quantities from their identically dimensioned non-angular quantity counterparts. Conversions to and from AngleKind quantities are supported through implementations of the From trait.

ConstituentConcentrationKind
Kind of constituent concentration in chemical mixtures, which separates mass concentration from mass density. This Kind is also applied to molar concentration and to catalytic activity concentration.

InformationKind
InformationKind is a Kind for separating information quantities from their identically dimensioned non-information quantity counterparts. Conversions to and from InformationKind quantities are supported through implementations of the From trait.

SolidAngleKind
SolidAngleKind is a Kind for separating quantities of solid angles from other identically dimensioned quantities. Conversions to and from SolidAngleKind quantities are supported through implementations of the From trait.

TemperatureKind
Kind of thermodynamic temperature.