Module uom::si::mass

Mass (base unit kilogram, kg).

Structs

attogram

Measurement unit.

carat

Measurement unit.

centigram

Measurement unit.

decagram

Measurement unit.

decigram

Measurement unit.

exagram

Measurement unit.

femtogram

Measurement unit.

gigagram

Measurement unit.

grain

Measurement unit.

gram

Measurement unit.

hectogram

Measurement unit.

hundredweight_long

Measurement unit.

hundredweight_short

Measurement unit.

kilogram

The kilogram is the SI unit of mass. It is defined by taking the fixed numerical value of the Planck constant h to be 6.626 070 15 × 10⁻³⁴ when expressed in the unit J s, which is equal to kg m² s⁻¹, where the meter and the second are defined in terms of c and ∆νCs.

megagram

Measurement unit.

microgram

Measurement unit.

milligram

Measurement unit.

nanogram

Measurement unit.

ounce

Measurement unit.

ounce_troy

Measurement unit.

pennyweight

Measurement unit.

petagram

Measurement unit.

picogram

Measurement unit.

pound

Measurement unit.

pound_troy

Measurement unit.

slug

Measurement unit.

teragram

Measurement unit.

ton

Measurement unit.

ton_assay

Measurement unit.

ton_long

Measurement unit.

ton_short

Measurement unit.

yoctogram

Measurement unit.

yottagram

Measurement unit.

zeptogram

Measurement unit.

zettagram

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Mass dimension, M (base unit kilogram, kg).

Mass

Mass (base unit kilogram, kg).