Module uom::si::mass_concentration

Mass concentration (base unit kilogram per cubic meter, kg · m⁻³).

Structs

attogram_per_cubic_meter
Measurement unit.

centigram_per_cubic_meter
Measurement unit.

decagram_per_cubic_meter
Measurement unit.

decigram_per_cubic_meter
Measurement unit.

exagram_per_cubic_meter
Measurement unit.

femtogram_per_cubic_decimeter
Measurement unit.

femtogram_per_cubic_meter
Measurement unit.

femtogram_per_deciliter
Measurement unit.

femtogram_per_liter
Measurement unit.

femtogram_per_milliliter
Measurement unit.

gigagram_per_cubic_meter
Measurement unit.

gram_per_cubic_decimeter
Measurement unit.

gram_per_cubic_meter
Measurement unit.

gram_per_deciliter
Measurement unit.

gram_per_liter
Measurement unit.

gram_per_milliliter
Measurement unit.

hectogram_per_cubic_meter
Measurement unit.

kilogram_per_cubic_decimeter
Measurement unit.

kilogram_per_cubic_meter
Measurement unit.

kilogram_per_deciliter
Measurement unit.

kilogram_per_liter
Measurement unit.

kilogram_per_milliliter
Measurement unit.

megagram_per_cubic_meter
Measurement unit.

microgram_per_cubic_decimeter
Measurement unit.

microgram_per_cubic_meter
Measurement unit.

microgram_per_deciliter
Measurement unit.

microgram_per_liter
Measurement unit.

microgram_per_milliliter
Measurement unit.

milligram_per_cubic_decimeter
Measurement unit.

milligram_per_cubic_meter
Measurement unit.

milligram_per_deciliter
Measurement unit.

milligram_per_liter
Measurement unit.

milligram_per_milliliter
Measurement unit.

nanogram_per_cubic_decimeter
Measurement unit.

nanogram_per_cubic_meter
Measurement unit.

nanogram_per_deciliter
Measurement unit.

nanogram_per_liter
Measurement unit.

nanogram_per_milliliter
Measurement unit.

petagram_per_cubic_meter
Measurement unit.

picogram_per_cubic_decimeter
Measurement unit.

picogram_per_cubic_meter
Measurement unit.

picogram_per_deciliter
Measurement unit.

picogram_per_liter
Measurement unit.

picogram_per_milliliter
Measurement unit.

teragram_per_cubic_meter
Measurement unit.

yoctogram_per_cubic_meter
Measurement unit.

yottagram_per_cubic_meter
Measurement unit.

zeptogram_per_cubic_meter
Measurement unit.

zettagram_per_cubic_meter
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of mass concentration, L⁻³M (base unit kilogram per cubic meter, kg · m⁻³).

MassConcentration
Mass concentration (base unit kilogram per cubic meter, kg · m⁻³).