Module uom::si::mass_density

Mass density (base unit kilogram per cubic meter, kg · m⁻³).

Structs

attogram_per_cubic_meter

Measurement unit.

carat_per_cubic_meter

Measurement unit.

centigram_per_cubic_meter

Measurement unit.

decagram_per_cubic_meter

Measurement unit.

decigram_per_cubic_meter

Measurement unit.

exagram_per_cubic_meter

Measurement unit.

femtogram_per_cubic_meter

Measurement unit.

gigagram_per_cubic_meter

Measurement unit.

grain_per_cubic_meter

Measurement unit.

grain_per_gallon

Measurement unit.

gram_per_cubic_centimeter

Measurement unit.

gram_per_cubic_meter

Measurement unit.

hectogram_per_cubic_meter

Measurement unit.

hundredweight_long_per_cubic_meter

Measurement unit.

hundredweight_short_per_cubic_meter

Measurement unit.

kilogram_per_cubic_meter

Derived unit of density.

megagram_per_cubic_meter

Measurement unit.

microgram_per_cubic_meter

Measurement unit.

milligram_per_cubic_meter

Measurement unit.

nanogram_per_cubic_meter

Measurement unit.

ounce_per_cubic_inch

Measurement unit.

ounce_per_cubic_meter

Measurement unit.

ounce_per_gallon

Measurement unit.

ounce_per_gallon_imperial

Measurement unit.

ounce_troy_per_cubic_meter

Measurement unit.

pennyweight_per_cubic_meter

Measurement unit.

petagram_per_cubic_meter

Measurement unit.

picogram_per_cubic_meter

Measurement unit.

pound_per_cubic_foot

Measurement unit.

pound_per_cubic_inch

Measurement unit.

pound_per_cubic_meter

Measurement unit.

pound_per_cubic_yard

Measurement unit.

pound_per_gallon

Measurement unit.

pound_per_gallon_imperial

Measurement unit.

pound_troy_per_cubic_meter

Measurement unit.

slug_per_cubic_foot

Measurement unit.

slug_per_cubic_meter

Measurement unit.

teragram_per_cubic_meter

Measurement unit.

ton_assay_per_cubic_meter

Measurement unit.

ton_long_per_cubic_meter

Measurement unit.

ton_long_per_cubic_yard

Measurement unit.

ton_per_cubic_meter

Measurement unit.

ton_short_per_cubic_meter

Measurement unit.

ton_short_per_cubic_yard

Measurement unit.

yoctogram_per_cubic_meter

Measurement unit.

yottagram_per_cubic_meter

Measurement unit.

zeptogram_per_cubic_meter

Measurement unit.

zettagram_per_cubic_meter

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

DensityDeprecated

Mass density (base unit kilogram per cubic meter, kg · m⁻³).

Dimension

Dimension of mass density, L⁻³M (base unit kilogram per cubic meter, kg · m⁻³).

MassDensity

Mass density (base unit kilogram per cubic meter, kg · m⁻³).