Module uom::si::mass_rate

Mass rate (base unit kilogram per second, kg · s⁻¹).

Structs

attogram_per_second
Measurement unit.

carat_per_second
Measurement unit.

centigram_per_second
Measurement unit.

decagram_per_second
Measurement unit.

decigram_per_second
Measurement unit.

exagram_per_second
Measurement unit.

femtogram_per_second
Measurement unit.

gigagram_per_second
Measurement unit.

grain_per_second
Measurement unit.

gram_per_day
Measurement unit.

gram_per_hour
Measurement unit.

gram_per_minute
Measurement unit.

gram_per_second
Measurement unit.

hectogram_per_second
Measurement unit.

hundredweight_long_per_second
Measurement unit.

hundredweight_short_per_second
Measurement unit.

kilogram_per_day
Measurement unit.

kilogram_per_hour
Measurement unit.

kilogram_per_minute
Measurement unit.

kilogram_per_second
Derived unit of mass rate.

megagram_per_second
Measurement unit.

microgram_per_second
Measurement unit.

milligram_per_second
Measurement unit.

nanogram_per_second
Measurement unit.

ounce_per_second
Measurement unit.

ounce_troy_per_second
Measurement unit.

pennyweight_per_second
Measurement unit.

petagram_per_second
Measurement unit.

picogram_per_second
Measurement unit.

pound_per_day
Measurement unit.

pound_per_hour
Measurement unit.

pound_per_minute
Measurement unit.

pound_per_second
Measurement unit.

pound_troy_per_second
Measurement unit.

slug_per_second
Measurement unit.

teragram_per_second
Measurement unit.

ton_assay_per_second
Measurement unit.

ton_long_per_second
Measurement unit.

ton_per_second
Measurement unit.

ton_short_per_hour
Measurement unit.

ton_short_per_second
Measurement unit.

yoctogram_per_second
Measurement unit.

yottagram_per_second
Measurement unit.

zeptogram_per_second
Measurement unit.

zettagram_per_second
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of mass rate, MT⁻¹ (base unit kilogram per second, kg · s⁻¹).

MassRate
Mass rate (base unit kilogram per second, kg · s⁻¹).