Module uom::si::molar_energy

Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

Structs

attojoule_per_mole
Measurement unit.

btu_39_per_mole
Measurement unit.

btu_59_per_mole
Measurement unit.

btu_60_per_mole
Measurement unit.

btu_it_per_mole
Measurement unit.

btu_per_mole
Measurement unit.

calorie_15_per_mole
Measurement unit.

calorie_20_per_mole
Measurement unit.

calorie_it_nutrition_per_mole
Measurement unit.

calorie_it_per_mole
Measurement unit.

calorie_nutrition_per_mole
Measurement unit.

calorie_per_mole
Measurement unit.

centijoule_per_mole
Measurement unit.

decajoule_per_mole
Measurement unit.

decawatt_hour_per_mole
Measurement unit.

decijoule_per_mole
Measurement unit.

electronvolt_per_mole
Measurement unit.

erg_per_mole
Measurement unit.

exajoule_per_mole
Measurement unit.

femtojoule_per_mole
Measurement unit.

foot_pound_force_per_mole
Measurement unit.

foot_poundal_per_mole
Measurement unit.

gigajoule_per_mole
Measurement unit.

gigawatt_hour_per_mole
Measurement unit.

hectojoule_per_mole
Measurement unit.

hectowatt_hour_per_mole
Measurement unit.

joule_per_mole
The derived unit of molar energy. Equivalent to kg · m²/(s² · mol).

kilocalorie_it_per_mole
Measurement unit.

kilocalorie_per_mole
Measurement unit.

kilogram_square_meter_per_second_squared_mole
Molar energy expressed in base units. Equivalent to J/mol.

kilojoule_per_mole
Measurement unit.

kilowatt_hour_per_mole
Measurement unit.

megajoule_per_mole
Measurement unit.

megawatt_hour_per_mole
Measurement unit.

microjoule_per_mole
Measurement unit.

microwatt_hour_per_mole
Measurement unit.

millijoule_per_mole
Measurement unit.

milliwatt_hour_per_mole
Measurement unit.

nanojoule_per_mole
Measurement unit.

petajoule_per_mole
Measurement unit.

petawatt_hour_per_mole
Measurement unit.

picojoule_per_mole
Measurement unit.

quad_per_mole
Measurement unit.

terajoule_per_mole
Measurement unit.

terawatt_hour_per_mole
Measurement unit.

therm_ec_per_mole
Measurement unit.

therm_us_per_mole
Measurement unit.

ton_tnt_per_mole
Measurement unit.

watt_hour_per_mole
Measurement unit.

watt_second_per_mole
Measurement unit.

yoctojoule_per_mole
Measurement unit.

yottajoule_per_mole
Measurement unit.

zeptojoule_per_mole
Measurement unit.

zettajoule_per_mole
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of molar energy, L²MT⁻²N⁻¹ (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).

MolarEnergy
Molar energy (base unit joule per mole, kg · m² · s⁻² · mol⁻¹).