Module uom::si::momentum

Momentum (base unit kilogram meter per second, kg · m · s⁻¹).

Structs

attogram_meter_per_second

Measurement unit.

centigram_meter_per_second

Measurement unit.

decagram_meter_per_second

Measurement unit.

decigram_meter_per_second

Measurement unit.

exagram_meter_per_second

Measurement unit.

femtogram_meter_per_second

Measurement unit.

gigagram_meter_per_second

Measurement unit.

gram_meter_per_second

Measurement unit.

hectogram_meter_per_second

Measurement unit.

kilogram_attometer_per_second

Measurement unit.

kilogram_centimeter_per_second

Measurement unit.

kilogram_decameter_per_second

Measurement unit.

kilogram_decimeter_per_second

Measurement unit.

kilogram_exameter_per_second

Measurement unit.

kilogram_femtometer_per_second

Measurement unit.

kilogram_gigameter_per_second

Measurement unit.

kilogram_hectometer_per_second

Measurement unit.

kilogram_kilometer_per_second

Measurement unit.

kilogram_megameter_per_second

Measurement unit.

kilogram_meter_per_day

Measurement unit.

kilogram_meter_per_hour

Measurement unit.

kilogram_meter_per_minute

Measurement unit.

kilogram_meter_per_second

Derived unit of momentum.

kilogram_micrometer_per_second

Measurement unit.

kilogram_millimeter_per_second

Measurement unit.

kilogram_nanometer_per_second

Measurement unit.

kilogram_petameter_per_second

Measurement unit.

kilogram_picometer_per_second

Measurement unit.

kilogram_terameter_per_second

Measurement unit.

kilogram_yoctometer_per_second

Measurement unit.

kilogram_yottameter_per_second

Measurement unit.

kilogram_zeptometer_per_second

Measurement unit.

kilogram_zettameter_per_second

Measurement unit.

megagram_meter_per_second

Measurement unit.

microgram_meter_per_second

Measurement unit.

milligram_meter_per_second

Measurement unit.

nanogram_meter_per_second

Measurement unit.

petagram_meter_per_second

Measurement unit.

picogram_meter_per_second

Measurement unit.

pound_foot_per_second

Measurement unit.

pound_inch_per_second

Measurement unit.

slug_foot_per_second

Measurement unit.

slug_inch_per_second

Measurement unit.

teragram_meter_per_second

Measurement unit.

ton_meter_per_second

Measurement unit.

yoctogram_meter_per_second

Measurement unit.

yottagram_meter_per_second

Measurement unit.

zeptogram_meter_per_second

Measurement unit.

zettagram_meter_per_second

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of momentum, LMT⁻¹ (base unit kilogram meter per second, kg · m · s⁻¹).

Momentum

Momentum (base unit kilogram meter per second, kg · m · s⁻¹).