Module uom::si::power

Power (base unit watt, m² · kg · s⁻³).

Structs

attowatt

Measurement unit.

centiwatt

Measurement unit.

decawatt

Measurement unit.

deciwatt

Measurement unit.

erg_per_second

Measurement unit.

exawatt

Measurement unit.

femtowatt

Measurement unit.

foot_pound_per_hour

Measurement unit.

foot_pound_per_minute

Measurement unit.

foot_pound_per_second

Measurement unit.

gigawatt

Measurement unit.

hectowatt

Measurement unit.

horsepower

Measurement unit.

horsepower_boiler

Measurement unit.

horsepower_electric

Measurement unit.

horsepower_imperial

Measurement unit.

horsepower_metric

Measurement unit.

hydraulic_horsepower

Measurement unit.

kilowatt

Measurement unit.

megawatt

Measurement unit.

microwatt

Measurement unit.

milliwatt

Measurement unit.

nanowatt

Measurement unit.

petawatt

Measurement unit.

picowatt

Measurement unit.

terawatt

Measurement unit.

watt

Derived unit of power.

yoctowatt

Measurement unit.

yottawatt

Measurement unit.

zeptowatt

Measurement unit.

zettawatt

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of power, L²MT⁻³ (base unit watt, m² · kg · s⁻³).

Power

Power (base unit watt, m² · kg · s⁻³).