Module uom::si::pressure

Pressure (base unit pascal, kg · m⁻¹ · s⁻²).

Structs

atmosphere
Measurement unit.

atmosphere_technical
Measurement unit.

attopascal
Measurement unit.

bar
Measurement unit.

centimeter_of_mercury
Measurement unit.

centimeter_of_mercury_0
Measurement unit.

centimeter_of_water
Measurement unit.

centimeter_of_water_4
Measurement unit.

centipascal
Measurement unit.

decapascal
Measurement unit.

decipascal
Measurement unit.

dyne_per_square_centimeter
Measurement unit.

exapascal
Measurement unit.

femtopascal
Measurement unit.

foot_of_mercury
Measurement unit.

foot_of_water
Measurement unit.

foot_of_water_39_2
Measurement unit.

gigapascal
Measurement unit.

gram_force_per_square_centimeter
Measurement unit.

hectopascal
Measurement unit.

inch_of_mercury
Measurement unit.

inch_of_mercury_32
Measurement unit.

inch_of_mercury_60
Measurement unit.

inch_of_water
Measurement unit.

inch_of_water_60
Measurement unit.

inch_of_water_39_2
Measurement unit.

kilogram_force_per_square_centimeter
Measurement unit.

kilogram_force_per_square_meter
Measurement unit.

kilogram_force_per_square_millimeter
Measurement unit.

kilopascal
Measurement unit.

kip_per_square_inch
Measurement unit.

megapascal
Measurement unit.

micropascal
Measurement unit.

millibar
Measurement unit.

millimeter_of_mercury
Measurement unit.

millimeter_of_water
Measurement unit.

millipascal
Measurement unit.

millitorr
Measurement unit.

nanopascal
Measurement unit.

pascal
Derived unit of pressure.

petapascal
Measurement unit.

picopascal
Measurement unit.

pound_force_per_square_foot
Measurement unit.

pound_force_per_square_inch
Measurement unit.

poundal_per_square_foot
Measurement unit.

psi
Measurement unit.

terapascal
Measurement unit.

torr
Measurement unit.

yoctopascal
Measurement unit.

yottapascal
Measurement unit.

zeptopascal
Measurement unit.

zettapascal
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of pressure, L⁻¹MT⁻² (base unit pascal, kg · m⁻¹ · s⁻²).

Pressure
Pressure (base unit pascal, kg · m⁻¹ · s⁻²).