Module uom::si::quantities

Quantity type aliases using the default base units and parameterized on the underlying storage type.

Type Definitions

Acceleration
Quantity type alias using the default base units parameterized on the underlying storage type.

AmountOfSubstance
Quantity type alias using the default base units parameterized on the underlying storage type.

Area
Quantity type alias using the default base units parameterized on the underlying storage type.

AvailableEnergy
Quantity type alias using the default base units parameterized on the underlying storage type.

Capacitance
Quantity type alias using the default base units parameterized on the underlying storage type.

Density
Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCharge
Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCurrent
Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricPotential
Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalConductance
Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalResistance
Quantity type alias using the default base units parameterized on the underlying storage type.

Energy
Quantity type alias using the default base units parameterized on the underlying storage type.

Force
Quantity type alias using the default base units parameterized on the underlying storage type.

Frequency
Quantity type alias using the default base units parameterized on the underlying storage type.

Inductance
Quantity type alias using the default base units parameterized on the underlying storage type.

Length
Quantity type alias using the default base units parameterized on the underlying storage type.

Luminance
Quantity type alias using the default base units parameterized on the underlying storage type.

LuminousIntensity
Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFlux
Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFluxDensity
Quantity type alias using the default base units parameterized on the underlying storage type.

Mass
Quantity type alias using the default base units parameterized on the underlying storage type.

MassRate
Quantity type alias using the default base units parameterized on the underlying storage type.

Power
Quantity type alias using the default base units parameterized on the underlying storage type.

Pressure
Quantity type alias using the default base units parameterized on the underlying storage type.

Ratio
Quantity type alias using the default base units parameterized on the underlying storage type.

TemperatureInterval
Quantity type alias using the default base units parameterized on the underlying storage type.

ThermodynamicTemperature
Quantity type alias using the default base units parameterized on the underlying storage type.

Time
Quantity type alias using the default base units parameterized on the underlying storage type.

Velocity
Quantity type alias using the default base units parameterized on the underlying storage type.

Volume
Quantity type alias using the default base units parameterized on the underlying storage type.

VolumeRate
Quantity type alias using the default base units parameterized on the underlying storage type.