Module uom::si::quantities

Quantity type aliases using the default base units and parameterized on the underlying storage type.

Type Definitions

Acceleration

Quantity type alias using the default base units parameterized on the underlying storage type.

AmountOfSubstance

Quantity type alias using the default base units parameterized on the underlying storage type.

Angle

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularAcceleration

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularJerk

Quantity type alias using the default base units parameterized on the underlying storage type.

AngularVelocity

Quantity type alias using the default base units parameterized on the underlying storage type.

Area

Quantity type alias using the default base units parameterized on the underlying storage type.

AvailableEnergy

Quantity type alias using the default base units parameterized on the underlying storage type.

Capacitance

Quantity type alias using the default base units parameterized on the underlying storage type.

CatalyticActivity

Quantity type alias using the default base units parameterized on the underlying storage type.

CatalyticActivityConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

Curvature

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCharge

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricCurrent

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricPotential

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalConductance

Quantity type alias using the default base units parameterized on the underlying storage type.

ElectricalResistance

Quantity type alias using the default base units parameterized on the underlying storage type.

Energy

Quantity type alias using the default base units parameterized on the underlying storage type.

Force

Quantity type alias using the default base units parameterized on the underlying storage type.

Frequency

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatFluxDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

HeatTransfer

Quantity type alias using the default base units parameterized on the underlying storage type.

Inductance

Quantity type alias using the default base units parameterized on the underlying storage type.

Information

Quantity type alias using the default base units parameterized on the underlying storage type.

InformationRate

Quantity type alias using the default base units parameterized on the underlying storage type.

Jerk

Quantity type alias using the default base units parameterized on the underlying storage type.

Length

Quantity type alias using the default base units parameterized on the underlying storage type.

Luminance

Quantity type alias using the default base units parameterized on the underlying storage type.

LuminousIntensity

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFlux

Quantity type alias using the default base units parameterized on the underlying storage type.

MagneticFluxDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

Mass

Quantity type alias using the default base units parameterized on the underlying storage type.

MassConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

MassDensity

Quantity type alias using the default base units parameterized on the underlying storage type.

MassRate

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarConcentration

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarEnergy

Quantity type alias using the default base units parameterized on the underlying storage type.

MolarMass

Quantity type alias using the default base units parameterized on the underlying storage type.

Momentum

Quantity type alias using the default base units parameterized on the underlying storage type.

Power

Quantity type alias using the default base units parameterized on the underlying storage type.

Pressure

Quantity type alias using the default base units parameterized on the underlying storage type.

RadiantExposure

Quantity type alias using the default base units parameterized on the underlying storage type.

Ratio

Quantity type alias using the default base units parameterized on the underlying storage type.

SolidAngle

Quantity type alias using the default base units parameterized on the underlying storage type.

SpecificHeatCapacity

Quantity type alias using the default base units parameterized on the underlying storage type.

TemperatureInterval

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermalConductivity

Quantity type alias using the default base units parameterized on the underlying storage type.

ThermodynamicTemperature

Quantity type alias using the default base units parameterized on the underlying storage type.

Time

Quantity type alias using the default base units parameterized on the underlying storage type.

Torque

Quantity type alias using the default base units parameterized on the underlying storage type.

Velocity

Quantity type alias using the default base units parameterized on the underlying storage type.

Volume

Quantity type alias using the default base units parameterized on the underlying storage type.

VolumeRate

Quantity type alias using the default base units parameterized on the underlying storage type.