Module uom::si::radiant_exposure

Radiant exposure (base unit joule per square meter, kg · s⁻²).

Structs

attojoule_per_square_meter

Measurement unit.

centijoule_per_square_centimeter

Measurement unit.

centijoule_per_square_meter

Measurement unit.

centijoule_per_square_millimeter

Measurement unit.

decajoule_per_square_centimeter

Measurement unit.

decajoule_per_square_meter

Measurement unit.

decajoule_per_square_millimeter

Measurement unit.

decijoule_per_square_centimeter

Measurement unit.

decijoule_per_square_meter

Measurement unit.

decijoule_per_square_millimeter

Measurement unit.

exajoule_per_square_meter

Measurement unit.

femtojoule_per_square_meter

Measurement unit.

gigajoule_per_square_centimeter

Measurement unit.

gigajoule_per_square_meter

Measurement unit.

gigajoule_per_square_millimeter

Measurement unit.

hectojoule_per_square_centimeter

Measurement unit.

hectojoule_per_square_meter

Measurement unit.

hectojoule_per_square_millimeter

Measurement unit.

joule_per_square_centimeter

Measurement unit.

joule_per_square_meter

Derived unit of radiant exposure.

joule_per_square_millimeter

Measurement unit.

kilojoule_per_square_centimeter

Measurement unit.

kilojoule_per_square_meter

Measurement unit.

kilojoule_per_square_millimeter

Measurement unit.

megajoule_per_square_centimeter

Measurement unit.

megajoule_per_square_meter

Measurement unit.

megajoule_per_square_millimeter

Measurement unit.

microjoule_per_square_centimeter

Measurement unit.

microjoule_per_square_meter

Measurement unit.

microjoule_per_square_millimeter

Measurement unit.

millijoule_per_square_centimeter

Measurement unit.

millijoule_per_square_meter

Measurement unit.

millijoule_per_square_millimeter

Measurement unit.

nanojoule_per_square_centimeter

Measurement unit.

nanojoule_per_square_meter

Measurement unit.

nanojoule_per_square_millimeter

Measurement unit.

petajoule_per_square_meter

Measurement unit.

picojoule_per_square_meter

Measurement unit.

terajoule_per_square_meter

Measurement unit.

yoctojoule_per_square_meter

Measurement unit.

yottajoule_per_square_meter

Measurement unit.

zeptojoule_per_square_meter

Measurement unit.

zettajoule_per_square_meter

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of radiant exposure, MT⁻² (base unit joule per square meter, kg · s⁻²).

RadiantExposure

Radiant exposure (base unit joule per square meter, kg · s⁻²).