Module uom::si::specific_heat_capacity

Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

This quantity might be used to define the heat capacity of a material. To define heat capacity of an object, use heat capacity.

Specific heat capacity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attojoule_per_kilogram_kelvin
Measurement unit.

btu_it_per_ounce_degree_fahrenheit
Measurement unit.

btu_it_per_pound_degree_fahrenheit
Measurement unit.

btu_it_per_ton_degree_fahrenheit
Measurement unit.

btu_per_ounce_degree_fahrenheit
Measurement unit.

btu_per_pound_degree_fahrenheit
Measurement unit.

btu_per_ton_degree_fahrenheit
Measurement unit.

centijoule_per_kilogram_kelvin
Measurement unit.

decajoule_per_kilogram_kelvin
Measurement unit.

decijoule_per_kilogram_kelvin
Measurement unit.

exajoule_per_kilogram_kelvin
Measurement unit.

femtojoule_per_kilogram_kelvin
Measurement unit.

gigajoule_per_kilogram_kelvin
Measurement unit.

hectojoule_per_kilogram_kelvin
Measurement unit.

joule_per_gram_degree_celsius
Measurement unit.

joule_per_kilogram_degree_celsius
Measurement unit.

joule_per_kilogram_kelvin
Derived unit of specific heat capacity expressed in derived units. Equivalent to m²/(s² · K).

kilojoule_per_gram_degree_celsius
Measurement unit.

kilojoule_per_kilogram_degree_celsius
Measurement unit.

kilojoule_per_kilogram_kelvin
Measurement unit.

megajoule_per_kilogram_kelvin
Measurement unit.

microjoule_per_kilogram_kelvin
Measurement unit.

millijoule_per_gram_degree_celsius
Measurement unit.

millijoule_per_kilogram_degree_celsius
Measurement unit.

millijoule_per_kilogram_kelvin
Measurement unit.

nanojoule_per_kilogram_kelvin
Measurement unit.

petajoule_per_kilogram_kelvin
Measurement unit.

picojoule_per_kilogram_kelvin
Measurement unit.

square_centimeter_per_second_squared_kelvin
Measurement unit.

square_kilometer_per_second_squared_kelvin
Measurement unit.

square_meter_per_second_squared_kelvin
The derived unit of specific heat capacity expressed in base units. Equivalent to J/(kg · K).

square_micrometer_per_second_squared_kelvin
Measurement unit.

square_millimeter_per_second_squared_kelvin
Measurement unit.

terajoule_per_kilogram_kelvin
Measurement unit.

yoctojoule_per_kilogram_kelvin
Measurement unit.

yottajoule_per_kilogram_kelvin
Measurement unit.

zeptojoule_per_kilogram_kelvin
Measurement unit.

zettajoule_per_kilogram_kelvin
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of specific heat capacity, L²T⁻²Th⁻¹ (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).

SpecificHeatCapacity
Specific heat capacity (base unit joule per kilogram kelvin, m² · s⁻² · K⁻¹).