Struct uom::si::Quantity

#[repr(transparent)]
pub struct Quantity<D: ?Sized, U: ?Sized, V> where
    D: Dimension,
    U: Units<V>,
    V: Num + Conversion<V>,  {
    pub dimension: PhantomData<D>,
    pub units: PhantomData<U>,
    pub value: V,
}

Property of a phenomenon, body or substance, where the property has a magnitude that can be expressed as a number and a reference.

The preferred method of creating a Quantity instance is to use the new constructor which is generic over the input unit and accepts the input value as it's only parameter.

// Create a length of 1 meter.
let l = Length::new::<meter>(1.0);

Quantity fields are public to allow for the creation of const values and instances of non-named Quantitys. This functionality will be deprecated and subsequently removed once the const fn feature is stabilized.

// Create a `const` length of 1 meter.
const L: Length = Length { dimension: PhantomData, units: PhantomData, value: 1.0, };
// Create a `const` area of 1 square meter explicitly without using the `Area` alias.
const A: Quantity<ISQ<P2, Z0, Z0, Z0, Z0, Z0, Z0>, SI<f32>, f32> =
    Quantity { dimension: PhantomData, units: PhantomData, value: 1.0, };

Using units for the wrong quantity will cause a compile error:

// error[E0277]: the trait bound `second: length::Unit` is not satisfied
let l = Length::new::<second>(1.0);

Mixing quantities will also cause a compile error:

// error[E0308]: mismatched types
let r = Length::new::<meter>(1.0) + Time::new::<second>(1.0);

// error[E0308]: mismatched types
let v: Velocity = Length::new::<meter>(1.0) * Time::new::<second>(1.0);

http://jcgm.bipm.org/vim/en/1.1.html

Generic Parameters

D: Quantity dimension. See Dimension.

U: Quantity base units. See Units.

V: Quantity value underlying storage type.

Fields
dimension: PhantomData<D>

Quantity dimension. See Dimension.

units: PhantomData<U>

Quantity base units. See Units.

value: V

Quantity value stored in the base units for the quantity.

Methods

impl<D: ?Sized, U: ?Sized, V> Quantity<D, U, V> where
D: Dimension,
U: Units<V>,
V: Num + Conversion<V>,

pub fn is_nan(self) -> bool where
V: Float,

Returns true if this value is NAN and false otherwise.

pub fn is_infinite(self) -> bool where
V: Float,

Returns true if this value is positive infinity or negative infinity and false otherwise.

pub fn is_finite(self) -> bool where
V: Float,

Returns true if this number is neither infinite nor NAN.

pub fn is_normal(self) -> bool where
V: Float,

Returns true if the number is neither zero, infinite, subnormal, or NAN.

pub fn classify(self) -> FpCategory where
V: Float,

Returns the floating point category of the number. If only one property is going to be tested, it is generally faster to use the specific predicate instead.

pub fn abs(self) -> Self where
V: Signed,

Computes the absolute value of self. Returns NAN if the quantity is NAN.

pub fn signum(self) -> Self where
V: Signed,

Returns a quantity that represents the sign of self.

1.0 of the base unit if the number is positive, +0.0, or INFINITY.

-1.0 of the base unit if the number is negative, -0.0, or NEG_INFINITY.

NAN if the number is NAN.

pub fn is_sign_positive(self) -> bool where
V: Float,

Returns true if self's sign bit is positive, including +0.0 and INFINITY.

pub fn is_sign_negative(self) -> bool where
V: Float,

Returns true if self's sign is negative, including -0.0 and NEG_INFINITY.

pub fn recip(
self
) -> Quantity<ISQ<Negate<D::L>, Negate<D::M>, Negate<D::T>, Negate<D::I>, Negate<D::Th>, Negate<D::N>, Negate<D::J>>, U, V> where
D::L: Neg,
D::M: Neg,
D::T: Neg,
D::I: Neg,
D::Th: Neg,
D::N: Neg,
D::J: Neg,
D::Kind: Div,
V: Float,

Takes the reciprocal (inverse) of a number, 1/x.

let f: Frequency = Time::new::<second>(1.0).recip();

pub fn powi<E>(
self,
e: E
) -> Quantity<ISQ<Prod<D::L, E>, Prod<D::M, E>, Prod<D::T, E>, Prod<D::I, E>, Prod<D::Th, E>, Prod<D::N, E>, Prod<D::J, E>>, U, V> where
D::L: Mul<E>,
D::M: Mul<E>,
D::T: Mul<E>,
D::I: Mul<E>,
D::Th: Mul<E>,
D::N: Mul<E>,
D::J: Mul<E>,
D::Kind: Mul,
E: Integer,
V: Pow<E, Output = V> + Conversion<V>,

Raises a quantity to an integer power.

let a: Area = Length::new::<meter>(3.0).powi(::uom::typenum::P2::new());

Generic Parameters

E: typenum::Integer power.

pub fn max(self, other: Self) -> Self where
V: Float,

Returns the maximum of the two quantities.

pub fn min(self, other: Self) -> Self where
V: Float,

Returns the minimum of the two quantities.