Module uom::si::temperature_interval

Temperature interval (base unit kelvin, K).

Temperature interval has the same dimensions as thermodynamic temperature but is not directly comparable. See thermodynamic temperature for a full explanation.

Structs

attokelvin
Measurement unit.

centikelvin
Measurement unit.

decakelvin
Measurement unit.

decikelvin
Measurement unit.

degree_celsius
Measurement unit.

degree_fahrenheit
Measurement unit.

degree_rankine
Measurement unit.

exakelvin
Measurement unit.

femtokelvin
Measurement unit.

gigakelvin
Measurement unit.

hectokelvin
Measurement unit.

kelvin
The kelvin is the SI unit of thermodynamic temperature. It is defined by taking the fixed numerical value of the Boltzmann constant k to be 1.380 649 × 10⁻²³ when expressed in the unit J K⁻¹, which is equal to kg m² s⁻² K⁻¹, where the kilogram, meter, and second are defined in terms of h, c, and ∆νCs.

kilokelvin
Measurement unit.

megakelvin
Measurement unit.

microkelvin
Measurement unit.

millikelvin
Measurement unit.

nanokelvin
Measurement unit.

petakelvin
Measurement unit.

picokelvin
Measurement unit.

terakelvin
Measurement unit.

yoctokelvin
Measurement unit.

yottakelvin
Measurement unit.

zeptokelvin
Measurement unit.

zettakelvin
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of temperature interval, Th (base unit kelvin, K).

TemperatureInterval
Temperature interval (base unit kelvin, K).