Module uom::si::thermal_conductivity

Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

Thermal conductivity has the same kind as temperature interval, as this quantity relates to change of temperature. Not of kind TemperatureKind, used by thermodynamic temperature. See thermodynamic temperature for a full explanation.

Structs

attogram_meter_per_second_cubed_kelvin
Measurement unit.

attowatt_per_meter_kelvin
Measurement unit.

centigram_meter_per_second_cubed_kelvin
Measurement unit.

centiwatt_per_meter_kelvin
Measurement unit.

decagram_meter_per_second_cubed_kelvin
Measurement unit.

decawatt_per_meter_kelvin
Measurement unit.

decigram_meter_per_second_cubed_kelvin
Measurement unit.

deciwatt_per_meter_kelvin
Measurement unit.

exagram_meter_per_second_cubed_kelvin
Measurement unit.

exawatt_per_meter_kelvin
Measurement unit.

femtogram_meter_per_second_cubed_kelvin
Measurement unit.

femtowatt_per_meter_kelvin
Measurement unit.

gigagram_meter_per_second_cubed_kelvin
Measurement unit.

gigawatt_per_meter_kelvin
Measurement unit.

gram_meter_per_second_cubed_kelvin
Measurement unit.

hectogram_meter_per_second_cubed_kelvin
Measurement unit.

hectowatt_per_meter_kelvin
Measurement unit.

kilogram_meter_per_second_cubed_degree_celsius
Measurement unit.

kilogram_meter_per_second_cubed_kelvin
Derived unit of thermal conductivity in base units. Equivalent to W/(m · K).

kilowatt_per_meter_degree_celsius
Measurement unit.

kilowatt_per_meter_kelvin
Measurement unit.

megagram_meter_per_second_cubed_kelvin
Measurement unit.

megawatt_per_meter_kelvin
Measurement unit.

microgram_meter_per_second_cubed_kelvin
Measurement unit.

microwatt_per_meter_kelvin
Measurement unit.

milligram_meter_per_second_cubed_kelvin
Measurement unit.

milliwatt_per_meter_degree_celsius
Measurement unit.

milliwatt_per_meter_kelvin
Measurement unit.

nanogram_meter_per_second_cubed_kelvin
Measurement unit.

nanowatt_per_meter_kelvin
Measurement unit.

petagram_meter_per_second_cubed_kelvin
Measurement unit.

petawatt_per_meter_kelvin
Measurement unit.

picogram_meter_per_second_cubed_kelvin
Measurement unit.

picowatt_per_meter_kelvin
Measurement unit.

teragram_meter_per_second_cubed_kelvin
Measurement unit.

terawatt_per_meter_kelvin
Measurement unit.

watt_per_meter_degree_celsius
Derived unit of thermal conductivity in derived units. Equivalent to kg · m/(s³ · K).

watt_per_meter_kelvin
Derived unit of thermal conductivity in derived units. Equivalent to kg · m/(s³ · K).

yoctogram_meter_per_second_cubed_kelvin
Measurement unit.

yoctowatt_per_meter_kelvin
Measurement unit.

yottagram_meter_per_second_cubed_kelvin
Measurement unit.

yottawatt_per_meter_kelvin
Measurement unit.

zeptogram_meter_per_second_cubed_kelvin
Measurement unit.

zeptowatt_per_meter_kelvin
Measurement unit.

zettagram_meter_per_second_cubed_kelvin
Measurement unit.

zettawatt_per_meter_kelvin
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of thermal conductivity, LMT⁻³Th⁻¹ (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).

ThermalConductivity
Thermal conductivity (base unit watt per meter kelvin, kg · m · s⁻³ · K⁻¹).