Module uom::si::thermodynamic_temperature

Thermodynamic temperature (base unit kelvin, K).

Thermodynamic temperature has the same dimensions as temperature interval but is not directly comparable. Thermodynamic temperature is the absolute measure of temperature and is one of the base quantities in the ISQ. Temperature interval is the measure of relative temperature difference between thermodynamic temperatures.

let tt = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let ti = TemperatureInterval::new::<temperature_interval::kelvin>(1.0);

// error[E0308]: mismatched types
let err = tt == ti;

Additionally, addition and subtraction are not implemented for thermodynamic temperature.

let t1 = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let t2 = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);

// error[E0308]: mismatched types
let err = t1 + t2;

A temperature interval can be added to or subtracted from a thermodynamic temperature.

let tt = ThermodynamicTemperature::new::<thermodynamic_temperature::kelvin>(1.0);
let ti = TemperatureInterval::new::<temperature_interval::kelvin>(1.0);

let result = tt + ti;

Structs

attokelvin
Measurement unit.

centikelvin
Measurement unit.

decakelvin
Measurement unit.

decikelvin
Measurement unit.

degree_celsius
Measurement unit.

degree_fahrenheit
Measurement unit.

degree_rankine
Measurement unit.

exakelvin
Measurement unit.

femtokelvin
Measurement unit.

gigakelvin
Measurement unit.

hectokelvin
Measurement unit.

kelvin
The kelvin is the SI unit of thermodynamic temperature. It is defined by taking the fixed numerical value of the Boltzmann constant k to be 1.380 649 × 10⁻²³ when expressed in the unit J K⁻¹, which is equal to kg m² s⁻² K⁻¹, where the kilogram, meter, and second are defined in terms of h, c, and ∆νCs.

kilokelvin
Measurement unit.

megakelvin
Measurement unit.

microkelvin
Measurement unit.

millikelvin
Measurement unit.

nanokelvin
Measurement unit.

petakelvin
Measurement unit.

picokelvin
Measurement unit.

terakelvin
Measurement unit.

yoctokelvin
Measurement unit.

yottakelvin
Measurement unit.

zeptokelvin
Measurement unit.

zettakelvin
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of thermodynamic temperature, Th (base unit kelvin, K).

ThermodynamicTemperature
Thermodynamic temperature (base unit kelvin, K).