Module uom::si::time

Time (base unit second, s1).

Structs

attosecond
Measurement unit.

centisecond
Measurement unit.

day
Measurement unit.

decasecond
Measurement unit.

decisecond
Measurement unit.

exasecond
Measurement unit.

femtosecond
Measurement unit.

gigasecond
Measurement unit.

hectosecond
Measurement unit.

hour
Measurement unit.

kilosecond
Measurement unit.

megasecond
Measurement unit.

microsecond
Measurement unit.

millisecond
Measurement unit.

minute
Measurement unit.

nanosecond
Measurement unit.

petasecond
Measurement unit.

picosecond
Measurement unit.

second
The second is the duration of 9 192 631 770 periods of the radiation corresponding to the transition between the two hyperfine levels of ground state of the caesium 133 atom.

shake
Measurement unit.

terasecond
Measurement unit.

year
Measurement unit.

yoctosecond
Measurement unit.

yottasecond
Measurement unit.

zeptosecond
Measurement unit.

zettasecond
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Time dimension, s1.

Time
Time (base unit second, s1).