Module uom::si::torque

Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).

Structs

attonewton_meter

Measurement unit.

centinewton_meter

Measurement unit.

decanewton_meter

Measurement unit.

decinewton_meter

Measurement unit.

dyne_centimeter

Measurement unit.

dyne_meter

Measurement unit.

exanewton_meter

Measurement unit.

femtonewton_meter

Measurement unit.

giganewton_meter

Measurement unit.

hectonewton_meter

Measurement unit.

kilogram_force_meter

Measurement unit.

kilonewton_meter

Measurement unit.

meganewton_meter

Measurement unit.

micronewton_meter

Measurement unit.

millinewton_meter

Measurement unit.

nanonewton_meter

Measurement unit.

newton_attometer

Measurement unit.

newton_centimeter

Measurement unit.

newton_decameter

Measurement unit.

newton_decimeter

Measurement unit.

newton_exameter

Measurement unit.

newton_femtometer

Measurement unit.

newton_gigameter

Measurement unit.

newton_hectometer

Measurement unit.

newton_kilometer

Measurement unit.

newton_megameter

Measurement unit.

newton_meter

Derived unit of torque.

newton_micrometer

Measurement unit.

newton_millimeter

Measurement unit.

newton_nanometer

Measurement unit.

newton_petameter

Measurement unit.

newton_picometer

Measurement unit.

newton_terameter

Measurement unit.

newton_yoctometer

Measurement unit.

newton_yottameter

Measurement unit.

newton_zeptometer

Measurement unit.

newton_zettameter

Measurement unit.

ounce_force_inch

Measurement unit.

petanewton_meter

Measurement unit.

piconewton_meter

Measurement unit.

pound_force_foot

Measurement unit.

pound_force_inch

Measurement unit.

teranewton_meter

Measurement unit.

yoctonewton_meter

Measurement unit.

yottanewton_meter

Measurement unit.

zeptonewton_meter

Measurement unit.

zettanewton_meter

Measurement unit.

Traits

Conversion

Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit

Marker trait to identify measurement units for the quantity. See Unit.

Functions

description

Quantity description.

Type Definitions

Dimension

Dimension of torque, L²MT⁻² (base unit newton meter, kg · m² · s⁻²).

Torque

Torque (aka moment of force) (base unit newton meter, kg · m² · s⁻²).