Trait uom::si::Dimension

pub trait Dimension: Send + Sync {
    type L: Integer;
    type M: Integer;
    type T: Integer;
    type I: Integer;
    type Th: Integer;
    type N: Integer;
    type J: Integer;
    type Kind: ?Sized;
}

Marker trait to express the dependence of a quantity on the base quantities of a system of quantities as a product of powers of factors corresponding to the base quantities, omitting any numerical factor.

http://jcgm.bipm.org/vim/en/1.7.html

Associated Types

type L: Integer

Length, one of the base quantities in the ISQ, denoted by the symbol L. The base unit for length is meter in the SI.

Quantity dimension.

type M: Integer

Mass, one of the base quantities in the ISQ, denoted by the symbol M. The base unit for mass is kilogram in the SI.

Quantity dimension.

type T: Integer

Time, one of the base quantities in the ISQ, denoted by the symbol T. The base unit for time is second in the SI.

Quantity dimension.

type I: Integer

Electric current, one of the base quantities in the ISQ, denoted by the symbol I. The base unit for electric current is ampere in the SI.

Quantity dimension.

type Th: Integer

Thermodynamic temperature, one of the base quantities in the ISQ, denoted by the symbol Th (Θ). The base unit for thermodynamic temperature is kelvin in the SI.

Quantity dimension.

type N: Integer

Amount of substance, one of the base quantities in the ISQ, denoted by the symbol N. The base unit for amount of substance is mole in the SI.

Quantity dimension.

type J: Integer

Luminous intensity, one of the base quantities in the ISQ, denoted by the symbol J. The base unit for luminous intensity is candela in the SI.

Quantity dimension.

type Kind: ?Sized

Kind of the quantity. Quantities of the same dimension but differing kinds are not comparable.