Trait uom::si::Units

pub trait Units<V>: Send + Sync where
    V: Conversion<V>, {
    type length: Unit + Conversion<V, T = V::T>;
    type mass: Unit + Conversion<V, T = V::T>;
    type time: Unit + Conversion<V, T = V::T>;
    type electric_current: Unit + Conversion<V, T = V::T>;
    type thermodynamic_temperature: Unit + Conversion<V, T = V::T>;
    type amount_of_substance: Unit + Conversion<V, T = V::T>;
    type luminous_intensity: Unit + Conversion<V, T = V::T>;
}

Marker trait to identify a system of units based on a set of base units of a system of quantities.

Generic Parameters

V: Underlying storage type trait is implemented for.

Associated Types

type length: Unit + Conversion<V, T = V::T>

Length, one of the base quantities in the ISQ, denoted by the symbol L. The base unit for length is meter in the SI.

Base unit.

type mass: Unit + Conversion<V, T = V::T>

Mass, one of the base quantities in the ISQ, denoted by the symbol M. The base unit for mass is kilogram in the SI.

Base unit.

type time: Unit + Conversion<V, T = V::T>

Time, one of the base quantities in the ISQ, denoted by the symbol T. The base unit for time is second in the SI.

Base unit.

type electric_current: Unit + Conversion<V, T = V::T>

Electric current, one of the base quantities in the ISQ, denoted by the symbol I. The base unit for electric current is ampere in the SI.

Base unit.

type thermodynamic_temperature: Unit + Conversion<V, T = V::T>

Thermodynamic temperature, one of the base quantities in the ISQ, denoted by the symbol Th (Θ). The base unit for thermodynamic temperature is kelvin in the SI.

Base unit.

type amount_of_substance: Unit + Conversion<V, T = V::T>

Amount of substance, one of the base quantities in the ISQ, denoted by the symbol N. The base unit for amount of substance is mole in the SI.

Base unit.

type luminous_intensity: Unit + Conversion<V, T = V::T>

Luminous intensity, one of the base quantities in the ISQ, denoted by the symbol J. The base unit for luminous intensity is candela in the SI.

Base unit.