Module uom::si::velocity

Velocity (base unit meter per second, m · s⁻¹).

Structs

attometer_per_second
Measurement unit.

centimeter_per_second
Measurement unit.

decameter_per_second
Measurement unit.

decimeter_per_second
Measurement unit.

exameter_per_second
Measurement unit.

femtometer_per_second
Measurement unit.

foot_per_hour
Measurement unit.

foot_per_minute
Measurement unit.

foot_per_second
Measurement unit.

gigameter_per_second
Measurement unit.

hectometer_per_second
Measurement unit.

inch_per_second
Measurement unit.

kilometer_per_hour
Measurement unit.

kilometer_per_second
Measurement unit.

knot
Measurement unit.

megameter_per_second
Measurement unit.

meter_per_second
Measurement unit.

micrometer_per_second
Measurement unit.

mile_per_hour
Measurement unit.

mile_per_minute
Measurement unit.

mile_per_second
Measurement unit.

millimeter_per_minute
Measurement unit.

millimeter_per_second
Measurement unit.

nanometer_per_second
Measurement unit.

petameter_per_second
Measurement unit.

picometer_per_second
Measurement unit.

terameter_per_second
Measurement unit.

yoctometer_per_second
Measurement unit.

yottameter_per_second
Measurement unit.

zeptometer_per_second
Measurement unit.

zettameter_per_second
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of velocity, LT⁻¹ (base unit meter per second, m · s⁻¹).

Velocity
Velocity (base unit meter per second, m · s⁻¹).