Module uom::si::volume

Volume (base unit cubic meter, m3).

Structs

acre_foot
Measurement unit.

barrel
Measurement unit.

bushel
Measurement unit.

cord
Measurement unit.

cubic_attometer
Measurement unit.

cubic_centimeter
Measurement unit.

cubic_decameter
Measurement unit.

cubic_decimeter
Measurement unit.

cubic_exameter
Measurement unit.

cubic_femtometer
Measurement unit.

cubic_foot
Measurement unit.

cubic_gigameter
Measurement unit.

cubic_hectometer
Measurement unit.

cubic_inch
Measurement unit.

cubic_kilometer
Measurement unit.

cubic_megameter
Measurement unit.

cubic_meter
Measurement unit.

cubic_micrometer
Measurement unit.

cubic_mile
Measurement unit.

cubic_millimeter
Measurement unit.

cubic_nanometer
Measurement unit.

cubic_petameter
Measurement unit.

cubic_picometer
Measurement unit.

cubic_terameter
Measurement unit.

cubic_yard
Measurement unit.

cubic_yoctometer
Measurement unit.

cubic_yottameter
Measurement unit.

cubic_zeptometer
Measurement unit.

cubic_zettameter
Measurement unit.

cup
Measurement unit.

fluid_ounce
Measurement unit.

fluid_ounce_imperial
Measurement unit.

gallon
Measurement unit.

gallon_imperial
Measurement unit.

gill
Measurement unit.

gill_imperial
Measurement unit.

liter
Measurement unit.

peck
Measurement unit.

pint_dry
Measurement unit.

pint_liquid
Measurement unit.

quart_dry
Measurement unit.

quart_liquid
Measurement unit.

register_ton
Measurement unit.

stere
Measurement unit.

tablespoon
Measurement unit.

teaspoon
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Volume dimension, m3.

Volume
Volume (base unit cubic meter, m3).