Module uom::si::volume_rate

Volume rate (base unit cubic meter per second, m³ · s⁻¹).

Structs

acre_foot_per_second
Measurement unit.

attoliter_per_second
Measurement unit.

barrel_per_second
Measurement unit.

bushel_per_second
Measurement unit.

centiliter_per_second
Measurement unit.

cord_per_second
Measurement unit.

cubic_attometer_per_second
Measurement unit.

cubic_centimeter_per_second
Measurement unit.

cubic_decameter_per_second
Measurement unit.

cubic_decimeter_per_second
Measurement unit.

cubic_exameter_per_second
Measurement unit.

cubic_femtometer_per_second
Measurement unit.

cubic_foot_per_minute
Measurement unit.

cubic_foot_per_second
Measurement unit.

cubic_gigameter_per_second
Measurement unit.

cubic_hectometer_per_second
Measurement unit.

cubic_inch_per_minute
Measurement unit.

cubic_inch_per_second
Measurement unit.

cubic_kilometer_per_second
Measurement unit.

cubic_megameter_per_second
Measurement unit.

cubic_meter_per_second
Measurement unit.

cubic_micrometer_per_second
Measurement unit.

cubic_mile_per_second
Measurement unit.

cubic_millimeter_per_second
Measurement unit.

cubic_nanometer_per_second
Measurement unit.

cubic_petameter_per_second
Measurement unit.

cubic_picometer_per_second
Measurement unit.

cubic_terameter_per_second
Measurement unit.

cubic_yard_per_minute
Measurement unit.

cubic_yard_per_second
Measurement unit.

cubic_yoctometer_per_second
Measurement unit.

cubic_yottameter_per_second
Measurement unit.

cubic_zeptometer_per_second
Measurement unit.

cubic_zettameter_per_second
Measurement unit.

cup_per_second
Measurement unit.

decaliter_per_second
Measurement unit.

deciliter_per_second
Measurement unit.

exaliter_per_second
Measurement unit.

femtoliter_per_second
Measurement unit.

fluid_ounce_imperial_per_second
Measurement unit.

fluid_ounce_per_second
Measurement unit.

gallon_imperial_per_second
Measurement unit.

gallon_per_day
Measurement unit.

gallon_per_minute
Measurement unit.

gallon_per_second
Measurement unit.

gigaliter_per_second
Measurement unit.

gill_imperial_per_second
Measurement unit.

gill_per_second
Measurement unit.

hectoliter_per_second
Measurement unit.

kiloliter_per_second
Measurement unit.

liter_per_second
Measurement unit.

megaliter_per_second
Measurement unit.

microliter_per_second
Measurement unit.

milliliter_per_second
Measurement unit.

nanoliter_per_second
Measurement unit.

peck_per_second
Measurement unit.

petaliter_per_second
Measurement unit.

picoliter_per_second
Measurement unit.

pint_dry_per_second
Measurement unit.

pint_liquid_per_second
Measurement unit.

quart_dry_per_second
Measurement unit.

quart_liquid_per_second
Measurement unit.

register_ton_per_second
Measurement unit.

stere_per_second
Measurement unit.

tablespoon_per_second
Measurement unit.

teaspoon_per_second
Measurement unit.

teraliter_per_second
Measurement unit.

yoctoliter_per_second
Measurement unit.

yottaliter_per_second
Measurement unit.

zeptoliter_per_second
Measurement unit.

zettaliter_per_second
Measurement unit.

Traits

Conversion
Trait to identify units which have a conversion factor for the Quantity. See Conversion<V>.

Unit
Marker trait to identify measurement units for the quantity. See Unit.

Functions

description
Quantity description.

Type Definitions

Dimension
Dimension of volume rate, L³T⁻¹ (base unit cubic meter per second, m³ · s⁻¹).

VolumeRate
Volume rate (base unit cubic meter per second, m³ · s⁻¹).